### 2024-03-06 #6447 @editorialbot
[PRE REVIEW]: aPhyloGeo: a multi-platform Python package for analyzing phylogenetic trees with climatic parameters
**Submitting author:** @TahiriNadia (<a href="http://orcid.org/0000-0002-1818-208X">Nadia Tahiri</a>)
**Repository:** https://github.com/tahiri-lab/aPhyloGeo
**Branch with paper.md** (empty if default branch): joss-journal
**Version:** v1.0.0
