### 2024-03-23 #6447 @fboehm
@jdalapicolla @AgustinPardo @annazhukova @fredericlemoine - would you be willing to review this submission for JOSS? JOSS reviews are checklist-driven. If you agree to review, you would need to assess whether the package meets each of a number of criteria. For those where it currently doesn't meet the criteria, we would ask that you offer suggestions to the authors on how to improve the existing submission. Thank you for considering this invitation!
