### 2024-03-26 58a9f77
Update paper.md
add Georges
