### 2024-03-27 fc960ff
Update paper.md adding affiliation 3
