### 2024-03-27 #6447 @AgustinPardo
> @jdalapicolla @AgustinPardo @annazhukova @fredericlemoine - would you be willing to review this submission for JOSS? JOSS reviews are checklist-driven. If you agree to review, you would need to assess whether the package meets each of a number of criteria. For those where it currently doesn't meet the criteria, we would ask that you offer suggestions to the authors on how to improve the existing submission. Thank you for considering this invitation!

Hello, thanks for taking me into account.
What is the review criteria of the assessment?
Out of curiosity, where did you get my username, where did you meet me?
