### 2024-03-28 #6447 @fboehm
@AgustinPardo thank you for the reply! My name is Fred, and I invited you to review at the suggestion of @TahiriNadia . I'm not sure how they know you or your work; I'll ask them to reply, too. The criteria, in general, involve assessing the quality of the software. There also is a very short manuscript that accompanies the software. For example, you would be asked to install the software package and see if it performs as advertised. If it contains unit tests, you would want to look at those.

I will try to remember to point you to the exact checklist tomorrow . I don't have it at my fingertips right now.

Thank you again for the consideration 😊
