### 2024-03-28 #6447 @fboehm
ALso, @AgustinPardo - it might be useful to examine other parts of the review thread that I linked to above, as it will give you a sense of how JOSS reviews often proceed. Unlike those of many journals, JOSS reviews are iterative and almost always lead, eventually, to publication. We typically don't reject a submission - instead, the reviewers offer suggestions to the authors to ensure that the publication criteria are attained.
