### 2024-03-28 #6447 @fboehm
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.25080/majora-212e5952-018 is OK
> - 10.25080/gerudo-f2bc6f59-00f is OK
>
> MISSING DOIs
>
> - 10.2307/2413638 may be a valid DOI for title: An alternating least squares approach to inferring...
> - 10.1093/molbev/msl072 may be a valid DOI for title: Neighbor-joining revealed
>
> INVALID DOIs
>
> - None
> ```

@TahiriNadia - please feel free to fix the dois, if possible.
