### 2024-03-29 30cb0ef
Update bib + DOI
