+++ paper/paper.bib
@@ -15,7 +15,8 @@
   number={1},
   pages={101--111},
   year={1997},
-  publisher={Society of Systematic Biologists}
+  publisher={Society of Systematic Biologists},
+  doi = {10.2307/2413638}
 }
 @InProceedings{ nadia_tahiri-proc-scipy-2023,
   author    = { {W}anlin {L}i and {N}adia {T}ahiri },
@@ -27,12 +28,24 @@
   doi       = { 10.25080/gerudo-f2bc6f59-00f }
 }
 @article{gascuel2006neighbor,
-  title={Neighbor-joining revealed},
-  author={Gascuel, Olivier and Steel, Mike},
-  journal={Molecular biology and evolution},
-  volume={23},
-  number={11},
-  pages={1997--2000},
-  year={2006},
-  publisher={Oxford University Press}
+  title = {Neighbor-joining revealed},
+  author = {Gascuel, Olivier and Steel, Mike},
+  journal = {Molecular biology and evolution},
+  volume = {23},
+  number = {11},
+  pages = {1997--2000},
+  year = {2006},
+  publisher = {Oxford University Press},
+  doi = {10.1093/molbev/msl072}
+}
+@article{robinson1981comparison,
+  title={Comparison of phylogenetic trees},
+  author={Robinson, David F and Foulds, Leslie R},
+  journal={Mathematical biosciences},
+  volume={53},
+  number={1-2},
+  pages={131--147},
+  year={1981},
+  publisher={Elsevier},
+  doi = 10.1016/0025-5564(81)90043-2}
 }
