+++ paper/paper.bib
@@ -47,5 +47,5 @@
   pages={131--147},
   year={1981},
   publisher={Elsevier},
-  doi = 10.1016/0025-5564(81)90043-2}
+  doi = {10.1016/0025-5564(81)90043-2}
 }
