### 2024-04-22 #6579 @mmore500
Some comments on the manuscript. Planning to follow up on the software content shortly.

## Introduction
There is a grammar issue in “between a genetic of species and its habitat during the reconstruction”

## Statement of Need
 — address in specific terms what specific scientific question(s) an be addressed through these analyses

## State of the Field
- by topological similarity, do you mean topological agreement?
- Consider phrasing in terms of first person (we) when discussing your own group

## figure
- in addition to the legend, it may make sense to additionally directly label the color background boxes (in addition to the symbol glyph labels, toes are nice!!)
- where possible, increase the font size, most of the figure can only be read by zooming way in.
- why are the biopython and python logos in the corner? if it’s meant as an acknowledgement, I think that would be better suited to other parts of the paper)
- what is a climate tree? This hasn’t yet been explicitly defined

## Pipeline:
- Can you clarify this sentence:  “…, forming the basis from which users obtain output data with essential calculations.” What data nd what calculations?
- In discussion of the figure, it’s unclear what “refer to the YAML file” means  —— that hasn’t yet been introduced
- rephrase “optimal” as “optimized”

## Multiprocessing:

What windows are you referring to?

## Dependencies:

Citations to the software would be appropriate.

## Conclusion:

- What specific problems or circumstances will the new methods allow to be tackled?
- “high standards in software development” —> “best practices in software development”
- Could the closing two sentences of the conclusion be made more specific and concrete? As written, they could describe a large number of software projects.

## Overall:

a specific application example or case study would greatly benefit the clarity of the manuscript
