### 2024-04-29 tahiri-lab/aPhyloGeo#40 @annazhukova
Add example usage with step-by-step instructions
Currently I have found a [Getting Started](https://github.com/tahiri-lab/aPhyloGeo/wiki/Getting-started) web page that gives some information on the pipeline use. Having an explicit step-by-step example with data provided in the github and commands stated would be very helpful to start using the software and understand how it is expected to work
