### 2024-04-29 tahiri-lab/aPhyloGeo#41 @annazhukova
Installation problem
I have tried to install aPhyloGeo using [these guidlines](https://github.com/tahiri-lab/aPhyloGeo/wiki/Installation-guide) on Ubuntu 22.04.4 LTS

I get the following issue:

```
$:~/Downloads/aPhyloGeo$ make
python aphylogeo/main.py
~/Downloads/aPhyloGeo/aphylogeo/main.py:1: DeprecationWarning:
Pyarrow will become a required dependency of pandas in the next major release of pandas (pandas 3.0),
(to allow more performant data types, such as the Arrow string type, and better interoperability with other libraries)
but was not found to be installed on your system.
If this would cause problems for you,
please provide us feedback at https://github.com/pandas-dev/pandas/issues/54466

  import pandas as pd
Traceback (most recent call last):
  File "~/Downloads/aPhyloGeo/aphylogeo/main.py", line 3, in <module>
    from aphylogeo.alignement import AlignSequences
ModuleNotFoundError: No module named 'aphylogeo'
make: *** [Makefile:2: aPhylogeo] Error 1
```
