### 2024-04-29 tahiri-lab/aPhyloGeo#42 @annazhukova
Our good first issue link in the contribution guidelines is broken
See [here](https://github.com/tahiri-lab/aPhyloGeo/wiki/Contribution-Guidelines)
