### 2024-05-09 fbed021
Add docs badge linking to GitHub wiki
