### 2024-05-09 tahiri-lab/aPhyloGeo#45 @mmore500
Link to Contribution Guidelines in GitHub Wiki appears broken 
The link "please follow our [contribution guidelines](https://github.com/tahiri-lab/aPhyloGeo/wiki/CONTRIBUTING.md)" does not take me to a contributing page; I stay on the same page.

This issue is part of JOSS Review https://github.com/openjournals/joss-reviews/issues/6579
