### 2024-05-09 tahiri-lab/aPhyloGeo#46 @mmore500
Contributing guidelines, how to run tests
Would it be possible to add some more detail about how to run tests to the contributing guidelines, or link to a page with instructions?

<https://github.com/tahiri-lab/aPhyloGeo/wiki/Contribution-Guidelines>: the current text only reads

> Ensure that your contributions pass all existing tests.
> If you add new features, include relevant tests.

part of review for JOSS https://github.com/openjournals/joss-reviews/issues/6579
