### 2024-05-09 tahiri-lab/aPhyloGeo#48 @mmore500
Statement of need could be clearer
As currently described on the README and the main page of the wiki, the only detail to the purpose/features of the library is "bioinformatics pipeline for phylogeography." This is a bit broad, and should be described more specifically and concretely.

JOSS review https://github.com/openjournals/joss-reviews/issues/6579
