### 2024-05-09 tahiri-lab/aPhyloGeo#49 @mmore500
Comments on Manuscript
Part of JOSS review https://github.com/openjournals/joss-reviews/issues/6579

typo/grammar: "between a genetic of species and its habitat during the reconstruction"

## Statement of Need
- address in specific terms what specific scientific question(s) an be addressed through these analyses

## State of the Field
- by topological similarity, do you mean topological agreement?
- consider rephrasing in first person voice, especially third-person references to "the Tahiri lab team"; the current third-person voice obscures the connection of what's being described to the present work.

## Figure
- In addition to the legend, it makes sense to directly label the color background boxes (in addition to the symbol glyph labels, those are nice!!)
- where possible, increase the font size, most of the figure can only be read by zooming way in.
- why are the biopython and python logos in the corner? if it’s meant as an acknowledgement, I think that would be better suited to other parts of the paper (i.e., citations in the text).
- what is a climate tree? This hasn’t yet been explicitly defined

## Pipeline:
- Can you clarify this sentence:  “…, forming the basis from which users obtain output data with essential calculations.” What data and what calculations?
- In discussion of the figure, it’s unclear what “refer to the YAML file” means  —— that hasn’t yet been introduced
- rephrase “optimal” as “optimized”

## Multiprocessing:
What windows are you referring to?

## Dependencies:
Citations to the software described would be appropriate.

## Conclusion:
- What specific problems or circumstances will the new methods allow to be tackled?
- “high standards in software development” —> “best practices in software development”
- Could the closing two sentences of the conclusion be made somewhat more specific?

## Overall:

As is, the reader would greatly benefit from attention to the concreteness/specificity of description of the library features and applications
A specific application example or case study would greatly benefit the clarity of the manuscript.
