### 2024-05-09 #6579 @mmore500
Apologies for the delay. I thought I had posted some comments on the manuscript much earlier, but the post must have not gone through. Luckily, I was able to find a copy of them and share them in https://github.com/tahiri-lab/aPhyloGeo/issues/49
