### 2024-05-09 #6579 @mmore500
I have updated my checklist, using `x` to mark complete items and `-` to mark items for which I have opened suggestions on the aPhyloGeo issue tracker. I'll keep an eye on these getting resolved, but please don't hesitate to @ me when it's time to take another look
