### 2024-05-13 tahiri-lab/aPhyloGeo#41 @geomarceau
Hi annazhukova,

We are using Poetry as packager and installer for dev environment. The main.py is mostly there to give some guidelines for developers on how to use the library.

We got some instructions for dev setup (that could be clarified) here: [https://github.com/tahiri-lab/aPhyloGeo/wiki/Contribution-Guidelines#development-setup](url)

1 - install poetry if you dont already have it. best practice is to install it with their setup: [curl -sSL https://install.python-poetry.org | python3 -](url) OR can be simply installed with pip install poetry.
2 - Clone the repository: git clone https://github.com/tahiri-lab/aPhyloGeo.git.
3 - Install dependencies: poetry install (poetry create and handle the Virtual Environment for you)
4 - Open the Venv: poetry shell

And you should get:

`C:\aphylogeo\aPhyloGeo>poetry install
Creating virtualenv aphylogeo-8DOfMLu--py3.11 in C:\Users\gmarceau\AppData\Local\pypoetry\Cache\virtualenvs
Installing dependencies from lock file

Package operations: 33 installs, 1 update, 0 removals

  - Installing certifi (2023.7.22)
  - Installing charset-normalizer (3.3.2)
  - Installing idna (3.4)
  - ....`

`C:\aphylogeo\aPhyloGeo>poetry shell
Spawning shell within C:\Users\gmarceau\AppData\Local\pypoetry\Cache\virtualenvs\aphylogeo-8DOfMLu--py3.11

C:\aphylogeo\aPhyloGeo>()

(aphylogeo-py3.11) C:\aphylogeo\aPhyloGeo>python aphylogeo/main.py

Starting alignement

Searching for the centroid
    Starting multiprocessing, this might take some time
     The first process is ran alone for calibration purposes

---
Available memory:  2.27 / 2.78 Gb
Active processes:  0  /  83408
....`

Hope it help.
