### 2024-05-28 7097fec
Update paper.md
corrected a grammar issue
