### 2024-05-28 fbf4248
Update paper.md
updated State of the Field - Advancements in Genomic Analysis
