+++ paper/paper.md
@@ -133,7 +133,7 @@ The *aPhyloGeo* pipeline serves as an integrative framework, bringing together a
 
 Looking ahead, *aPhyloGeo* aims to integrate new functionalities, including clustering techniques based on similarity derived from multiple sequence alignments and a more computationally efficient alignment methodology. The incorporation of novel metrics, such as the Quartet metric and bipartition, aims to provide users with improved insight for making nuanced decisions regarding their datasets through a comprehensive assessment of genetic diversity.
 
-Adhering strictly to best practices in software development, this research not only seeks to provide immediate solutions but also aims to position *aPhyloGeo* as a reliable and adaptable platform. Striving to contribute meaningfully to the field of phylogeographic analysis, the pipeline is committed to offering users a sophisticated suite of tools that seamlessly adapt to the evolving landscape of genetic research. Through these improvements, the pipeline aims to make a valuable and enduring contribution to the scientific community, enhancing the standards of reproducibility and usability.
+By adhering to best practices in software development and embracing open-source principles, *aPhyloGeo* not only provides a reliable and adaptable platform for current phylogeographic research but also lays the groundwork for future expansion and innovation. The ongoing integration of cutting-edge methodologies, such as clustering based on multiple sequence alignments and optimized alignment algorithms, demonstrates our commitment to continually enhance the capabilities of aPhyloGeo. Moreover, the incorporation of novel metrics like the Quartet metric and bipartition analysis empowers users with enhanced decision-making tools for assessing genetic diversity and its relationship to environmental factors. Through these targeted improvements, aPhyloGeo aims to foster a new era of reproducible, accessible, and comprehensive phylogeographic analysis, ultimately deepening our understanding of the complex interactions between species and their environments.
 
 # Acknowledgements
 
