+++ paper/paper.md
@@ -38,7 +38,7 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
-The cross-platform application for phylogenetic tree analysis with climate parameters, *aPhyloGeo*, is a robust pipeline designed for comprehensive phylogenetic analyses using genetic and climate data. This Python API, available on [PyPI](https://pypi.org/project/aphylogeo/), offers a suite of analyses tailored to various scenarios, enabling the examination of datasets at three distinct levels: 1) genetic, 2) climatic, and 3) biogeography correlation, all within a unified package. Similarity at these levels, evaluated through metrics such as least squares distance [@felsenstein1997alternating], Euclidean distance, and Robinson-Foulds distance [@robinson1981comparison], significantly influences the assumptions guiding the identification of correlations between the genetic makeup of a species and its habitat during the reconstruction of the multiple alignment necessary for phylogenetic inference [@gascuel2006neighbor].
+The cross-platform application for phylogenetic tree analysis with climate parameters, *aPhyloGeo*, is a robust pipeline designed for comprehensive phylogenetic analyses using genetic and climate data. This Python API, available at [PyPI] (https://pypi.org/project/aphylogeo/), offers a suite of analyses adapted to different scenarios, enabling the examination of datasets at three distinct levels: 1) genetic, 2) climatic and 3) biogeographic correlation, all within a unified package. Similarity at these levels, assessed by metrics such as least squares distance [@felsenstein1997alternating], Euclidean distance and Robinson-Foulds distance [@robinson1981comparison], significantly influences the assumptions guiding the identification of correlations between species genetics and their habitats when reconstructing the multiple alignment required for phylogenetic inference [@gascuel2006neighbor].
 
 By utilizing the *aPhyloGeo* Python API, users can programmatically implement sophisticated phylogenetic analyses without the need for a graphical interface. This API provides a powerful and flexible toolset for conducting analyses, allowing users to tailor the application to their specific research needs. Through this approach, *aPhyloGeo* facilitates a nuanced understanding of the interplay between genetic evolution and environmental factors in the context of species adaptation, all within the Python programming environment.
 
