+++ paper/paper.md
@@ -38,11 +38,16 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
-The cross-platform application for phylogenetic tree analysis with climate parameters, *aPhyloGeo*, is a robust pipeline designed for comprehensive phylogenetic analyses using genetic and climate data. This Python API, available at [PyPI] (https://pypi.org/project/aphylogeo/), offers a suite of analyses adapted to different scenarios, enabling the examination of datasets at three distinct levels: 1) genetic, 2) climatic and 3) biogeographic correlation, all within a unified package. Similarity at these levels, assessed by metrics such as least squares distance [@felsenstein1997alternating], Euclidean distance and Robinson-Foulds distance [@robinson1981comparison], significantly influences the assumptions guiding the identification of correlations between species genetics and their habitats when reconstructing the multiple alignment required for phylogenetic inference [@gascuel2006neighbor].
+*aPhyloGeo*, a versatile and open-source Python application available at [PyPI] (https://pypi.org/project/aphylogeo/), is designed to elucidate the complex relationship between species evolution and environmental pressures, with a particular focus on climate. By integrating genetic and climatic data, *aPhyloGeo* empowers researchers to investigate the mechanisms of evolutionary adaptation and pinpoint genetic regions potentially influenced by environmental factors.
 
-By utilizing the *aPhyloGeo* Python API, users can programmatically implement sophisticated phylogenetic analyses without the need for a graphical interface. This API provides a powerful and flexible toolset for conducting analyses, allowing users to tailor the application to their specific research needs. Through this approach, *aPhyloGeo* facilitates a nuanced understanding of the interplay between genetic evolution and environmental factors in the context of species adaptation, all within the Python programming environment.
+The software's core strength lies in its comprehensive phylogenetic analysis pipeline, encompassing three distinct levels of investigation: genetic relationships, climatic impact assessment, and biogeographic correlations. This multi-faceted approach facilitates a holistic understanding of how species evolve and adapt to their environments. For example, researchers studying the genetic basis of high-altitude adaptation in birds could utilize aPhyloGeo to construct phylogenetic trees from genetic data, analyze oxygen levels across different altitudes, and identify correlations between specific genes and hypoxic conditions.
 
-By selecting an appropriate gene list for the available data defined on a set of species to explain the adaptation of the species according to the Darwinian hypothesis, the user can be confident that these assumptions are taken into account in *aPhyloGeo*.
+In another scenario, scientists investigating the impact of climate change on marine biodiversity could employ *aPhyloGeo* to examine the genetic diversity of coral species, assess changes in sea surface temperatures over time, and pinpoint genetic markers associated with thermal tolerance.
+These examples demonstrate the wide range of research questions that *aPhyloGeo* can address, making it an invaluable tool for evolutionary biologists, ecologists, and conservationists alike.
+
+Underlying *aPhyloGeo*'s analyses are robust algorithms employing metrics such as least squares distance [@felsenstein1997alternating] , Euclidean distance, and Robinson-Foulds distance [@robinson1981comparison] to quantify similarity across different levels. This rigorous approach ensures that the identification of correlations is statistically sound, while adhering to the principles of phylogenetic inference [@gascuel2006neighbor]. The software's modular design and Python interface offer flexibility, allowing users to tailor analyses to their specific research questions and datasets. Additionally, *aPhyloGeo*'s open-source nature fosters collaboration and transparency within the scientific community.
+
[…]
