+++ paper/paper.md
@@ -78,7 +78,7 @@ Navigating the *aPhyloGeo* workflow (refer to \autoref{fig:figure1}) is indispen
 
 The diagram below illustrates the workflow of the algorithm, consisting of several key blocks, each highlighted with a distinct color.
 
-- **First Block (Light Blue):** This block creates climate trees based on input climate data and validates the input parameters using a YAML file. This involves processing climatic variables such as temperature, precipitation, and elevation to construct phylogenetic trees that represent the relationships between geographic locations based on their climatic similarity.
+- **First Block (Light Blue):** This block creates climate trees based on input climate data (CSV file) and validates the input parameters using a YAML file. More precisely, the climate trees were generated by calculating the pairwise differences between each value of the species' habitats, normalized between the minimum and maximum of the parameter. This process resulted in a symmetric square matrix. From this matrix, the climate tree was inferred using the Neighbor-Joining method. This involves processing climatic variables such as temperature, precipitation, and elevation to construct phylogenetic trees that represent the relationships between geographic locations based on their climatic similarity.
 
 - **Second Block (Light Green):** This block creates phylogenetic trees based on input genetic data and performs input parameter validation (refer to the [YAML file](../aphylogeo/params.yaml)). This entails aligning DNA or amino acid sequences, inferring phylogenetic relationships using various methods (e.g., maximum likelihood, Bayesian inference), and assessing the statistical support for the inferred tree topology.
 
