+++ paper/paper.md
@@ -135,9 +135,9 @@ This comprehensive methodology ensures a nuanced and high-quality analysis, cont
 
 The *aPhyloGeo* pipeline serves as an integrative framework, bringing together a variety of advanced analytical methodologies for diverse datasets, covering both genetic and climatic aspects. By consolidating these analyses within a unified platform, users can simplify their exploration of different tools while ensuring greater reproducibility in research outcomes.
 
-Looking ahead, *aPhyloGeo* aims to integrate new functionalities, including clustering techniques based on similarity derived from multiple sequence alignments and a more computationally efficient alignment methodology. The incorporation of novel metrics, such as the Quartet metric and bipartition, aims to provide users with improved insight for making nuanced decisions regarding their datasets through a comprehensive assessment of genetic diversity.
+Looking ahead, *aPhyloGeo* aims to integrate new functionalities, including clustering techniques based on similarity derived from multiple sequence alignments and a more computationally efficient alignment methodology. The incorporation of novel metrics, such as the Euclidean distance and Robinson-Foulds distance, aims to provide users with improved insight for making nuanced decisions regarding their datasets through a comprehensive assessment of genetic diversity.
 
-By adhering to best practices in software development and embracing open-source principles, *aPhyloGeo* not only provides a reliable and adaptable platform for current phylogeographic research but also lays the groundwork for future expansion and innovation. The ongoing integration of cutting-edge methodologies, such as clustering based on multiple sequence alignments and optimized alignment algorithms, demonstrates our commitment to continually enhance the capabilities of aPhyloGeo. Moreover, the incorporation of novel metrics like the Quartet metric and bipartition analysis empowers users with enhanced decision-making tools for assessing genetic diversity and its relationship to environmental factors. Through these targeted improvements, aPhyloGeo aims to foster a new era of reproducible, accessible, and comprehensive phylogeographic analysis, ultimately deepening our understanding of the complex interactions between species and their environments.
+By adhering to best practices in software development and embracing open-source principles, *aPhyloGeo* not only provides a reliable and adaptable platform for current phylogeographic research but also lays the groundwork for future expansion and innovation. The ongoing integration of cutting-edge methodologies, such as clustering based on multiple sequence alignments and optimized alignment algorithms, demonstrates our commitment to continually enhance the capabilities of aPhyloGeo. Moreover, the incorporation of novel metrics like the Euclidean distance and Robinson-Foulds distance empowers users with enhanced decision-making tools for assessing genetic diversity and its relationship to environmental factors. Through these targeted improvements, aPhyloGeo aims to foster a new era of reproducible, accessible, and comprehensive phylogeographic analysis, ultimately deepening our understanding of the complex interactions between species and their environments.
 
 # Acknowledgements
 
