### 2024-06-03 #6579 @hazem-dev
> Some comments on the manuscript. Planning to follow up on the software content shortly.
>
> ## Introduction
> There is a grammar issue in “between a genetic of species and its habitat during the reconstruction”
>
> ## Statement of Need
> Address in more specific terms what specific scientific question(s) an be addressed through these analyses.
>
> ## State of the Field
> * by topological similarity, do you mean topological agreement?
> * Consider phrasing in terms of first person (we) when discussing your own group
>
> ## figure
> * in addition to the legend, it may make sense to additionally directly label the color background boxes (in addition to the symbol glyph labels, toes are nice!!)
> * where possible, increase the font size, most of the figure can only be read by zooming way in.
> * why are the biopython and python logos in the corner? if it’s meant as an acknowledgement, I think that would be better suited to other parts of the paper)
> * what is a climate tree? This hasn’t yet been explicitly defined
>
> ## Pipeline:
> * Can you clarify this sentence:  “…, forming the basis from which users obtain output data with essential calculations.” What data nd what calculations?
> * In discussion of the figure, it’s unclear what “refer to the YAML file” means  —— that hasn’t yet been introduced
> * rephrase “optimal” as “optimized”
>
> ## Multiprocessing:
> What windows are you referring to?
>
> ## Dependencies:
> Citations to the software would be appropriate.
>
> ## Conclusion:
> * What specific problems or circumstances will the new methods allow to be tackled?
> * “high standards in software development” —> “best practices in software development”
> * Could the closing two sentences of the conclusion be made more specific and concrete? As written, they could describe a large number of software projects.
>
> ## Overall:
> a specific application example or case study would greatly benefit the clarity of the manuscript

@mmore500, thank you for your insightful review and constructive feedback. Your comments were instrumental in guiding the revisions we made to the manuscript.
Here's a summary of how we've addressed your specific concerns:

1. 	**Introduction and Statement of Need:** We've sharpened the focus and articulated the core scientific questions our research aims to answer.
2. 	**State of the Field:** The terminology related to "topological agreement" has been refined for greater clarity and precision.
3. 	**Figure and Pipeline:** We've improved the figure with clearer labels and enhanced readability. Explanations of each pipeline step and the use of the YAML file have been added for a deeper understanding of the process.
4. 	**Multiprocessing, Dependencies, and Conclusion:** We've incorporated your feedback on window types, software citations, and refined the concluding remarks for a stronger finish.

In addition, we emphasize the real-world applications of aPhyloGeo in several key areas:

- 	**Phylogeography:** Tracing the geographic distribution and evolutionary relationships of organisms, such as tracking the spread of invasive species or uncovering historical migration patterns.
- 	**Conservation Biology:** Assessing the impact of climate change on species distributions and identifying critical areas for conservation efforts.
- 	**Epidemiology:** Tracking the origin and spread of infectious diseases, pinpointing potential reservoirs, and forecasting future outbreaks.

To further showcase the practical utility of aPhyloGeo, we're actively using it in our ongoing project, [iphylogeo++](https://github.com/tahiri-lab/iPhyloGeo_plus_plus), available at GitHub.
We sincerely appreciate your time and expertise. Please don't hesitate to share any additional thoughts or suggestions you may have.
