+++ paper/paper.bib
@@ -49,3 +49,150 @@
   publisher={Elsevier},
   doi = {10.1016/0025-5564(81)90043-2}
 }
+@article{katoh2013mafft,
+  title={MAFFT multiple sequence alignment software version 7: improvements in performance and usability},
+  author={Katoh, Kazutaka and Standley, Daron M},
+  journal={Molecular biology and evolution},
+  volume={30},
+  number={4},
+  pages={772--780},
+  year={2013},
+  publisher={Society for Molecular Biology and Evolution}
+}
+@article{cock2009biopython,
+  title={Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+  author={Cock, Peter JA and Antao, Tiago and Chang, Jeffrey T and Chapman, Brad A and Cox, Cymon J and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and others},
+  journal={Bioinformatics},
+  volume={25},
+  number={11},
+  pages={1422},
+  year={2009},
+  publisher={Oxford University Press}
+}
+@article{huerta2016ete,
+  title={ETE 3: reconstruction, analysis, and visualization of phylogenomic data},
+  author={Huerta-Cepas, Jaime and Serra, Fran{\c{c}}ois and Bork, Peer},
+  journal={Molecular biology and evolution},
+  volume={33},
+  number={6},
+  pages={1635--1638},
+  year={2016},
+  publisher={Society for Molecular Biology and Evolution}
+}
+@article{sukumaran2010dendropy,
+  title={DendroPy: a Python library for phylogenetic computing},
+  author={Sukumaran, Jeet and Holder, Mark T},
+  journal={Bioinformatics},
+  volume={26},
+  number={12},
+  pages={1569--1571},
+  year={2010},
+  publisher={Oxford University Press}
+}
+@article{cornish2021biopython,
+  title={A Biopython-based method for comprehensively searching for eponyms in Pubmed},
+  author={Cornish, Toby C and Kricka, Larry J and Park, Jason Y},
+  journal={MethodsX},
+  volume={8},
+  pages={101264},
+  year={2021},
+  publisher={Elsevier}
+}
+@article{hung2016sequence,
+  title={Sequence alignment and homology search with BLAST and ClustalW},
+  author={Hung, Jui-Hung and Weng, Zhiping},
+  journal={Cold Spring Harbor Protocols},
+  volume={2016},
+  number={11},
+  pages={pdb--prot093088},
+  year={2016},
+  publisher={Cold Spring Harbor Laboratory Press}
+}
+@article{edgar2004muscle,
+  title={MUSCLE: a multiple sequence alignment method with reduced time and space complexity},
+  author={Edgar, Robert C},
+  journal={BMC bioinformatics},
+  volume={5},
+  pages={1--19},
+  year={2004},
+  publisher={Springer}
+}
+@article{li2018minimap2,
+  title={Minimap2: pairwise alignment for nucleotide sequences},
+  author={Li, Heng},
+  journal={Bioinformatics},
+  volume={34},
+  number={18},
+  pages={3094--3100},
+  year={2018},
+  publisher={Oxford University Press}
+}
+@inproceedings{labib2019hamming,
+  title={Hamming distance completeness},
+  author={Labib, Karim and Uznanski, Przemyslaw and Wolleb-Graf, Daniel},
+  booktitle={30th Annual Symposium on Combinatorial Pattern Matching (CPM 2019)},
+  volume={128},
+  pages={14},
+  year={2019},
+  organization={Schloss Dagstuhl-Leibniz-Zentrum f{\"u}r Informatik}
+}
+@article{yujian2007normalized,
[…]
