+++ paper/paper.md
@@ -68,7 +68,7 @@ By maintaining the highest standards, this research aims to make a significant c
 # State of the Field - Advancements in Genomic Analysis
 The field of genomic analysis has progressed significantly in recent years, notably in the creation of tools and algorithms to explore the intricate relationship between genetic variation and environmental factors. The Tahiri lab team's innovative 2021 algorithm for identifying sub-sequences within genes [@nadia_tahiri-proc-scipy-2022], and its subsequent application to SARS-CoV-2 data in 2023 [@nadia_tahiri-proc-scipy-2023], stand out as substantial contributions to this field, enhancing our comprehension of the genetic underpinnings of adaptation across various species and environments.
 
-In the broader field of phylogeography, substantial methodological advancements have also occurred. Several Python packages provide functionalities pertinent to phylogeographic analysis, but often in a fragmented way. Biopython, a cornerstone in bioinformatics, excels at handling genetic sequences and basic phylogenetic tasks, yet falls short in integrating environmental data. DendroPy, a robust library for phylogenetic trees, aids in visualizing phylogeographic patterns but requires additional tools for comprehensive analysis. While SciPy's statistical prowess could be harnessed for custom analyses, its complexity demands a strong background in statistical programming. GeoPandas, adept at handling geospatial data, is useful for mapping genetic or environmental distributions, but lacks seamless integration with genetic data analysis tools. In summary, while powerful individual tools exist, a comprehensive and user-friendly Python package specifically designed for phylogeographic analysis remains a gap to be filled.
+In the broader field of phylogeography, substantial methodological advancements have also occurred. Several Python packages provide functionalities pertinent to phylogeographic analysis, but often in a fragmented way. Biopython [@cornish2021biopython], a cornerstone in bioinformatics, excels at handling genetic sequences and basic phylogenetic tasks, yet falls short in integrating environmental data. DendroPy, a robust library for phylogenetic trees, aids in visualizing phylogeographic patterns but requires additional tools for comprehensive analysis. While SciPy's statistical prowess could be harnessed for custom analyses, its complexity demands a strong background in statistical programming. GeoPandas, adept at handling geospatial data, is useful for mapping genetic or environmental distributions, but lacks seamless integration with genetic data analysis tools. In summary, while powerful individual tools exist, a comprehensive and user-friendly Python package specifically designed for phylogeographic analysis remains a gap to be filled.
 
 Statistical approaches, including generalized linear models (GLMs) and mixed models, are increasingly used to investigate the relationship between genetic variation and environmental variables. These methods enable researchers to quantify the relative influence of various factors, such as climate, geography, and demography, on observed patterns of genetic diversity.
 
@@ -100,10 +100,10 @@ The algorithm supports multiprocessing, allowing the simultaneous analysis of mu
 
 This work relies on the following main software packages:
 
-- [ete3](https://pypi.org/project/ete3/) version 3.1.3 (Huerta-Cepas, Serra & Bork, 2016), available under the GNU General Public License (GPL) (GPLv3), is used for phylogenetic tree manipulation and visualization.
-- [Bio](https://pypi.org/project/bio/) version 1.5.9 (Cock et al., 2009), available under the New BSD License, provides a wide range of bioinformatics tools and functionalities.
[…]
