+++ aphylogeo/requirements.txt
@@ -1,1 +1,1 @@
-aphylogeo
+aphylogeo==1.0.0
