### 2024-06-03 tahiri-lab/aPhyloGeo#40 @hazem-dev
Thank you for bringing this to our attention. We have taken steps to improve the documentation and user support for our software. Specifically, we have updated the [wiki](https://github.com/tahiri-lab/aPhyloGeo/wiki) to offer a more detailed and clear explanation of the software's functionality. This should help users understand the capabilities and features more comprehensively.

Furthermore, we have added a thorough [tutorial](https://github.com/tahiri-lab/aPhyloGeo/wiki/Tutorial) designed to guide users through the analysis process step-by-step. This tutorial includes detailed instructions and examples to ensure that users can effectively utilize the software for their specific needs.
