### 2024-06-03 #6579 @hazem-dev
> I have filled in my checklist, and here are a few comments:
>
> > * [ ]   **Contribution and authorship:** Has the submitting author (@TahiriNadia) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> I have not ticked this iteam as looking at the [contributors page](https://github.com/tahiri-lab/aPhyloGeo/graphs/contributors) I saw that the user my-linh-luu seemed to have contributed substantially to the software but does not seem to be on the authors’ list
>
> > * [ ]   **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> From what I understood from the guidlines, the software needs to be already quite established (cited, used) which does not yet seem to be the case here.
>
> > Some factors that may be considered by editors and reviewers when judging effort include:
> > Age of software (is this a well-established software project) / length of commit history
>
> I saw that the first tag was in June 2022, so one would expect a quite established software with quite some usages
>
> > Number of commits:
>
> I have counted about 600 on the [contributors page](https://github.com/tahiri-lab/aPhyloGeo/graphs/contributors)
>
> > Number of authors:
>
> 3 (on the paper, more on github)
>
> > Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.
>
> I have assessed the LOC with `wc -l *.py` command on the aphylogeo folder: 2152
>
> > Whether the software has already been cited in academic papers.
>
> According to google-scholar the only citation is a [self-citation](https://scholar.google.ca/citations?view_op=view_citation&hl=en&user=cOfpZhEAAAAJ&sortby=pubdate&citation_for_view=cOfpZhEAAAAJ:tKAzc9rXhukC)
>
> > Whether the software is sufficiently useful that it is likely to be cited by your peer group.
>
> I think that a clear example of an analysis pipeline with the software would highly increase the chances of future citations (as people would know how to use the software for their data).
>
> > ### Functionality
> >
> > * []  **Installation:** Does installation proceed as outlined in the documentation?
>
> See [this issue](https://github.com/tahiri-lab/aPhyloGeo/issues/41)
>
> > * [ ]   **Functionality:** Have the functional claims of the software been confirmed?
>
> I haven't managed to install it (see above)
>
> > * [ ]   **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> I haven't managed to install it (see above)
>
> > ### Documentation
>
> > * [ ]   **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> See [issue](https://github.com/tahiri-lab/aPhyloGeo/issues/40)
>
> > * [ ]   **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> See [issue](https://github.com/tahiri-lab/aPhyloGeo/issues/40)
>
> > * [ x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> There are guidlines but they could be better illustrated: see [issue](https://github.com/tahiri-lab/aPhyloGeo/issues/42)
>
> > ### Software paper
> >
> > * [ ]   **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
> From what I understood at first the goal of the software is to allow to analyse the correlation between the climate and species evolution. Though reading further and especially looking at the figure, it seems to me that the goal might be to select gene regions that have the most correlation (?) Overall I think a use case, an example of a data analysis with aPhyloGeo in the article (and the corresponding data and code available and described in GitHub) would highly facilitate understanding the goals.
>
> > * [ ]   **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
>
> There is the corresponding section, but it reads to me a bit too general. A more concrete example would help here too.
>
> > * [ ]   **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
> The authors only mention their own previous work in this section. I would expect here seeing what can be done with other packages, for instance used in classical phylogeography: Ancestral Character Reconstruction for geographic and or climate characters, [GLM](https://doi.org/10.1371/journal.ppat.1003932) with climate as a factor etc.
>
> > * [ ]   **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
>
> Adding more information to the State of the field section would add more references too

Thank you for your thorough review and valuable feedback. We appreciate you taking the time to evaluate aPhyloGeo and providing your insights.
We understand your concerns regarding authorship and contributions, and we want to assure you that all contributors have been duly recognized.
The author list and contribution statement have been updated to accurately reflect their involvement.
Good remark, regarding software maturity, we acknowledge that aPhyloGeo was registered in 2022. However, it has undergone significant development since then, culminating in the current v1.0 release, which we consider a stable and reliable version.
We have also expanded the documentation and included more examples to enhance user experience.
We acknowledge the limited number of citations for aPhyloGeo, which we attribute to the software's relative novelty. We are actively promoting the software and are confident that its usage and citations will increase as more researchers become aware of its capabilities.
Good observation! Indeed, there are 600 commits listed on the contributors' page and over 2000 lines of code (LOC) in the GitHub repository.
Regarding the citation, we have updated the citation to 19 references including two of our lab.
In response to your suggestions, we have updated the wiki to clarify the software's functionality.
Additionally, we already have added a tutorial in to guide users through the analysis process step-by-step.
We believe that the changes we have made effectively address the concerns raised in your review.
We are confident that aPhyloGeo is a valuable tool for researchers investigating the relationship between phylogenetic trees and climatic parameters.
Thank you again for your valuable contribution.
