### 2024-06-03 tahiri-lab/aPhyloGeo#45 @hazem-dev
Thank you for bringing this to our attention. We have fixed the issue with the "please follow our contribution guidelines" link. It now correctly redirects to the contributing page.
