### 2024-06-04 aaad9b3
script python3
