### 2024-06-04 tahiri-lab/aPhyloGeo#46 @hazem-dev
Thank you for your suggestion. We have updated the [Contribution Guidelines](https://github.com/tahiri-lab/aPhyloGeo/wiki/Contribution-Guidelines) to include detailed instructions on how to run tests. You can find the new section under **Testing**.
