+++ paper/paper.bib
@@ -57,7 +57,8 @@
   number={4},
   pages={772--780},
   year={2013},
-  publisher={Society for Molecular Biology and Evolution}
+  publisher={Society for Molecular Biology and Evolution},
+  doi = {10.1093/molbev/mst010},
 }
 @article{cock2009biopython,
   title={Biopython: freely available Python tools for computational molecular biology and bioinformatics},
