+++ paper/paper.bib
@@ -68,7 +68,8 @@
   number={11},
   pages={1422},
   year={2009},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btp163},
 }
 @article{huerta2016ete,
   title={ETE 3: reconstruction, analysis, and visualization of phylogenomic data},
