+++ paper/paper.bib
@@ -79,7 +79,8 @@
   number={6},
   pages={1635--1638},
   year={2016},
-  publisher={Society for Molecular Biology and Evolution}
+  publisher={Society for Molecular Biology and Evolution},
+  doi = {10.1093/molbev/msw046},
 }
 @article{sukumaran2010dendropy,
   title={DendroPy: a Python library for phylogenetic computing},
