+++ paper/paper.bib
@@ -90,7 +90,8 @@
   number={12},
   pages={1569--1571},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btq228},
 }
 @article{cornish2021biopython,
   title={A Biopython-based method for comprehensively searching for eponyms in Pubmed},
