+++ paper/paper.bib
@@ -100,7 +100,8 @@
   volume={8},
   pages={101264},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.mex.2021.101264},
 }
 @article{hung2016sequence,
   title={Sequence alignment and homology search with BLAST and ClustalW},
