+++ paper/paper.bib
@@ -111,7 +111,8 @@
   number={11},
   pages={pdb--prot093088},
   year={2016},
-  publisher={Cold Spring Harbor Laboratory Press}
+  publisher={Cold Spring Harbor Laboratory Press},
+  doi = {10.1101/pdb.prot093088},
 }
 @article{edgar2004muscle,
   title={MUSCLE: a multiple sequence alignment method with reduced time and space complexity},
