+++ paper/paper.bib
@@ -121,7 +121,8 @@
   volume={5},
   pages={1--19},
   year={2004},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1186/1471-2105-5-113}
 }
 @article{li2018minimap2,
   title={Minimap2: pairwise alignment for nucleotide sequences},
@@ -131,7 +132,8 @@
   number={18},
   pages={3094--3100},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/bty191}
 }
 @inproceedings{labib2019hamming,
   title={Hamming distance completeness},
@@ -140,7 +142,8 @@
   volume={128},
   pages={14},
   year={2019},
-  organization={Schloss Dagstuhl-Leibniz-Zentrum f{\"u}r Informatik}
+  organization={Schloss Dagstuhl-Leibniz-Zentrum f{\"u}r Informatik},
+  doi = {10.4230/LIPIcs.CPM.2019.14}
 }
 @article{yujian2007normalized,
   title={A normalized Levenshtein distance metric},
@@ -150,7 +153,8 @@
   number={6},
   pages={1091--1095},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/TPAMI.2007.1078}
 }
 @article{zhao2019string,
   title={String correction using the Damerau-Levenshtein distance},
@@ -159,7 +163,8 @@
   volume={20},
   pages={1--28},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1186/s12859-019-2819-0}
 }
 @article{pradhan2015review,
   title={A Review on Text Similarity Technique used in IR and its Application},
@@ -169,7 +174,8 @@
   number={9},
   pages={29--34},
   year={2015},
-  publisher={Foundation of Computer Science}
+  publisher={Foundation of Computer Science},
+  doi = {10.5120/21257-4109}
 }
 @article{waterman1978similarity,
   title={On the similarity of dendrograms},
@@ -179,7 +185,8 @@
   number={4},
   pages={789--800},
   year={1978},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/0022-5193(78)90137-6}
 }
 @article{bag2019efficient,
   title={An efficient recommendation generation using relevant Jaccard similarity},
@@ -188,7 +195,8 @@
   volume={483},
   pages={53--64},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.ins.2019.01.023}
 }
 @article{li2020generic,
   title={Generic sao similarity measure via extended s{\o}rensen-dice index},
@@ -197,8 +205,6 @@
   volume={8},
   pages={66538--66552},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/ACCESS.2020.2984024}
 }
-
-
-
