+++ paper/paper.md
@@ -42,14 +42,14 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Abstract
-*aPhyloGeo*, a versatile and open-source Python application available at [PyPI](https://pypi.org/project/aphylogeo/), is designed to elucidate the complex relationship between species evolution and environmental pressures, with a particular focus on climate. By integrating genetic and climatic data, *aPhyloGeo* empowers researchers to investigate the mechanisms of evolutionary adaptation and pinpoint genetic regions potentially influenced by environmental factors.
+*aPhyloGeo* is a Python library designed to elucidate the complex relationship between species evolution and environmental pressures, with a particular focus on climate. By integrating genetic and climatic data, *aPhyloGeo* empowers researchers to investigate the mechanisms of evolutionary adaptation and pinpoint genetic regions potentially influenced by environmental factors.
 
 The software's core strength lies in its comprehensive phylogenetic analysis pipeline, encompassing three distinct levels of investigation: genetic relationships, climatic impact assessment, and biogeographic correlations. This multi-faceted approach facilitates a holistic understanding of how species evolve and adapt to their environments. For example, researchers studying the genetic basis of high-altitude adaptation in birds could utilize aPhyloGeo to construct phylogenetic trees from genetic data, analyze oxygen levels across different altitudes, and identify correlations between specific genes and hypoxic conditions.
 
 In another scenario, scientists investigating the impact of climate change on marine biodiversity could employ *aPhyloGeo* to examine the genetic diversity of coral species, assess changes in sea surface temperatures over time, and pinpoint genetic markers associated with thermal tolerance.
-These examples demonstrate the wide range of research questions that *aPhyloGeo* can address, making it an invaluable tool for evolutionary biologists, ecologists, and conservationists alike.
+These examples demonstrate the wide range of research questions that *aPhyloGeo* can address, making it an valuable tool for evolutionary biologists, ecologists, and conservationists alike.
 
-Underlying *aPhyloGeo*'s analyses are robust algorithms employing metrics such as least squares distance [@felsenstein1997alternating], Euclidean distance, and Robinson-Foulds distance [@robinson1981comparison] to quantify similarity across different levels. This rigorous approach ensures that the identification of correlations is statistically sound, while adhering to the principles of phylogenetic inference [@gascuel2006neighbor]. The software's modular design and Python interface offer flexibility, allowing users to tailor analyses to their specific research questions and datasets. Additionally, *aPhyloGeo*'s open-source nature fosters collaboration and transparency within the scientific community.
+Underlying *aPhyloGeo*'s analyses are robust algorithms employing metrics such as least squares distance [@felsenstein1997alternating], Euclidean distance, and Robinson-Foulds distance [@robinson1981comparison] to quantify similarity across different levels. This approach ensures that the identification of correlations is statistically sound, while adhering to the principles of phylogenetic inference [@gascuel2006neighbor]. The software's modular design and Python interface offer flexibility, allowing users to tailor analyses to their specific research questions and datasets. Additionally, *aPhyloGeo*'s open-source nature fosters collaboration and transparency within the scientific community.
 
 By enabling researchers to explore the complex interplay between genetics and environment, *aPhyloGeo* contributes to a deeper understanding of evolutionary processes. This knowledge not only enhances our appreciation of the natural world but also informs conservation efforts in the face of climate change and other environmental challenges. By identifying genetic adaptations to changing environments, *aPhyloGeo* can help prioritize species and populations for conservation, ultimately contributing to the preservation of biodiversity on our planet.
 
@@ -57,28 +57,29 @@ By enabling researchers to explore the complex interplay between genetics and en
 
[…]
