### 2024-06-08 tahiri-lab/aPhyloGeo#49 @mmore500
Some further comments on the manuscript:

- I appreciate the details that have been clarified in the current draft of the manuscript. Some further revisions be necessary to meet (or approximately meet) the JOSS word count limit of 1,000 words. The current word count is 2,400 words. There are some opportunities to tighten/prune the language and remove detail that is redundant (appears in more than one place) or less important.
- “State of the Field” should contain references that are non-self citations, or should be retitled.
- Several of my concerns remain regarding the figure. Most importantly, the text on the figure is too small to be legible without zooming in significantly. I would also recommend removing the Python and BioPython logos, so as to not imply direct affiliation with these projects.
- What is the taxonomic unit of “genetic trees” referred to in the text? Are these gene trees (i.e., taxonomic unit is individual genes), or is the taxonomic unit populations or species?
- With respect to the “third block,” what is gained by constructing “climate trees” and doing tree distance comparisons as opposed to directly testing for correlation between climate characteristics and genetic similarity? Relatedly, are potentially-confounding spatial effects  controlled for in current analyses; if so, how?
    - i.e., organisms in nearby spatial areas tend to be more related due to mobility/migration limitations and climate variables in nearby spatial areas tend to be similar due to geographic similarity. I suspect some level of correlation would be expected in the absence of selection effects?
- Several cited dependencies do not appear in `setup.py`. Are these dependencies necessary for end users? Either way, end users will not be pinned to specific versions so version information may be omitted or should be indicated >=.

Smaller comments:
- On line 28, it is somewhat vague what “across different levels” is intended to refer to.
- On line 134, it is unclear what the “workflow below” refers to. Is this supposed to refer to the section on multiprocessing?
- in “Pipeline” section, revise for consistent verb tense (some sentences are in past tense)
- iphylogeo++ could be mentioned in the manuscript
- on line 113, would it be clearer to say, `(refer to the project’s example YAML file)` —- I would also recommend linking to GitHub content (e.g., the YAML file) at a commit or a tag so that the link stays valid in the future
