### 2024-06-08 #6579 @mmore500
Thank you to the authors for their revisions to the manuscript. They have been helpful to providing a clearer understanding of the software and underlying methods. I have raised some further comments in https://github.com/tahiri-lab/aPhyloGeo/issues/49#issuecomment-2156182763

One concern is the current word count. As I mentioned in the review, I believe that there is potential for good space savings with some focused editing for brevity in how things are phrased, etc.

I plan to follow up in testing the software on my own machine shortly.
