### 2024-06-08 baf9acc
reduce state of fields
