### 2024-06-08 a030c56
reduce conclusion
