+++ paper/paper.md
@@ -110,11 +110,8 @@ This comprehensive methodology ensures a nuanced and high-quality analysis, cont
 
 # Conclusion
 
-The *aPhyloGeo* pipeline serves as an integrative framework, bringing together a variety of advanced analytical methodologies for diverse datasets, covering both genetic and climatic aspects. By consolidating these analyses within a unified platform, users can simplify their exploration of different tools while ensuring greater reproducibility in research outcomes.
+The \textit{aPhyloGeo} pipeline provides an integrative framework for advanced analysis of genetic and climatic data, promoting simplicity and reproducibility. Future enhancements include clustering techniques and improved alignment methods, alongside novel metrics like Quarted and bipartition distances for better genetic diversity assessment. Emphasizing software development best practices and open-source principles (e.g., iPhyloGeo++), \textit{aPhyloGeo} ensures reliability and sets the stage for ongoing innovation. By prioritizing user empowerment and comprehensive analysis, it aims to advance phylogeographic research and deepen our understanding of species-environment interactions.
 
-Looking ahead, *aPhyloGeo* aims to integrate new functionalities, including clustering techniques based on similarity derived from multiple sequence alignments and a more computationally efficient alignment methodology. The incorporation of novel metrics, such as the Euclidean distance and Robinson-Foulds distance, aims to provide users with improved insight regarding their datasets through a comprehensive assessment of genetic diversity.
-
-By adhering to best practices in software development and embracing open-source principles, *aPhyloGeo* not only provides a reliable and adaptable platform for current phylogeographic research but also lays the groundwork for future expansion and innovation. The ongoing integration of cutting-edge methodologies, such as clustering based on multiple sequence alignments and optimized alignment algorithms, demonstrates our commitment to continually enhance the capabilities of aPhyloGeo. Moreover, the incorporation of novel metrics like the Euclidean distance and Robinson-Foulds distance empowers users with enhanced decision-making tools for assessing genetic diversity and its relationship to environmental factors. Through these targeted improvements, aPhyloGeo aims to foster reproducible, accessible, and comprehensive phylogeographic analysis, ultimately deepening our understanding of the complex interactions between species and their environments.
 
 # Acknowledgements
 
