### 2024-06-11 tahiri-lab/aPhyloGeo#40 @annazhukova
Thank you for the tutorial, it is really nice and I could reproduce the steps !

There seems to be a little problem in the beginning:

It starts like...
> We selected only 5 of 38 lineages with regional characteristics for further study (see Koshkarov et al., 2022).
There seems to be a bit of context missing. I am not sure what are those 38 lineages.

> As shown in Figure 2...
I am not sure where to look for the Figure 2

And also it would be really great if you could add a few words on how to interpret the results obtained by the pipeline, "geneticTrees.json" for example.
