### 2024-06-11 tahiri-lab/aPhyloGeo#41 @annazhukova
Hi, thank you for adding the development setup instructions.

I managed to run them all right, however with a few minor adjustments that should probably be listed on the page:

1. It seems like python 3.9 is needed (worth specifying in theinstructions). When I tried with python 3.8 `poetry install` produced an issue "Current Python version (3.8.19) is not allowed by the project (>=3.9.0)."
2. Step 2 should be change the directory to aPhyloGeo: `cd aPhyloGeo`
3. If one uses conda `poetry shell` step is not needed (maybe also worth mentioning)
4. Step 4 currently states `python.exe aphylogeo/main.py`. It is maybe good for Windows, but for linux should be `python aphylogeo/main.py` (again would not hurt to mention)

Thanks!
