### 2024-06-11 #6579 @annazhukova
I would like to thank the authors for the updates. I have ticked several new checkpoint. There are however still things missing in my opinion.

In particular:

> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There is a tutotrial now (thank you!) which is very easy to reproduce, however what is missing are explanations on what exactly it does and why and how to interpret the results (see [this issue](https://github.com/tahiri-lab/aPhyloGeo/issues/40), which you closed). The same holds for the article itself: there is nowhere an explanation nor example on how to use real world data with aPhyloGeo and what would be the results of the analysis and how to interpret them.

> ### Software paper
> * [ ]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

See above

> * [ ]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

Formally there is a "statement of need" section however its content "Climat change and human activities are rapidly impacting biodiversity and population dynamics. To understand these complexities, biologists use phylogeography, which examines the interplay between genetics, geography, and past/present climate." does not explain at all how aPhyloGeo comes into play here, and what it brings. Nor what are the alternatives (other work).

> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

I could not see any other phylogeographic/climatic package being mentioned. The only other software descriptions are descriptions of tools being used as part of aPhyloGeo pipeline.

(I am actually worndering if the [paper I am looking at](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06579/joss.06579/10.21105.joss.06579.pdf) is the good one and whether there is something else with more explanations and information?)
