### 2024-06-11 tahiri-lab/aPhyloGeo#40 @hazem-dev
Thank you, @annazhukova. Your feedback was instrumental in identifying areas for improvement in our documentation. We have updated the wiki to provide clearer context regarding the selection of the 5 lineages from the original 38 and to offer detailed guidance on interpreting the "geneticTrees.json" file. These revisions significantly enhance the clarity and usability of our GitHub repository's wiki.
