+++ paper/paper.bib
@@ -208,3 +208,30 @@
   publisher={IEEE},
   doi = {10.1109/ACCESS.2020.2984024}
 }
+@article{drummond2007beast,
+  title={BEAST: Bayesian evolutionary analysis by sampling trees},
+  author={Drummond, Alexei J and Rambaut, Andrew},
+  journal={BMC evolutionary biology},
+  volume={7},
+  pages={1--8},
+  year={2007},
+  publisher={Springer}
+}
+@article{ronquist2012mrbayes,
+  title={MrBayes 3.2: efficient Bayesian phylogenetic inference and model choice across a large model space},
+  author={Ronquist, Fredrik and Teslenko, Maxim and Van Der Mark, Paul and Ayres, Daniel L and Darling, Aaron and H{\"o}hna, Sebastian and Larget, Bret and Liu, Liang and Suchard, Marc A and Huelsenbeck, John P},
+  journal={Systematic biology},
+  volume={61},
+  number={3},
+  pages={539--542},
+  year={2012},
+  publisher={Oxford University Press}
+}
+@incollection{scott2009spatial,
+  title={Spatial statistics in ArcGIS},
+  author={Scott, Lauren M and Janikas, Mark V},
+  booktitle={Handbook of applied spatial analysis: Software tools, methods and applications},
+  pages={27--41},
+  year={2009},
+  publisher={Springer}
+}
