### 2024-06-11 #6579 @hazem-dev
Thank you, @annazhukova, for your thorough review of our paper and for providing such detailed and constructive feedback. We greatly appreciate the time and effort you dedicated to this process. Your comments have been invaluable in guiding us to substantially revise and enhance our manuscript.

In response to your feedback, we have implemented several changes to address the issues you raised.  Specifically, we have:

*   **Example Usage:** Expanded the tutorial with step-by-step explanations and interpretation guidelines to ensure clarity and understanding for all users.
*   **Software Paper Summary:**  Refined the description to better convey the high-level functionality and purpose of the software to a wider audience, including potential applications in various analysis scenarios.
*   **Statement of Need:** Thoroughly detailed how aPhyloGeo directly addresses the challenges posed by climate change and human activities on biodiversity and population dynamics. Additionally, we have clarified the unique contributions of aPhyloGeo compared to existing tools.
*   **State of the Field:** Provided a comprehensive comparison of aPhyloGeo with other widely used phylogeographic and climatic packages, highlighting its unique strengths and advantages.

We apologize for any misunderstanding regarding the initial closure of the comment.  Our intention was not to dismiss your valuable feedback. We closed the issue prematurely after implementing initial solutions, believing the matter was resolved. However, we have since reopened the issue and diligently worked to address all of your concerns in this revised version of the manuscript.

We sincerely hope that these revisions meet your expectations and would be grateful for your continued feedback.  We value your expertise and believe your input is essential for improving the quality and impact of our work.

Thank you, @annazhukova and @fboehm, again for your time and consideration.
