### 2024-06-11 ffbac77
update methods section
