### 2024-06-11 tahiri-lab/aPhyloGeo#49 @hazem-dev
Thank you, @mmore500, for your detailed and insightful feedback. We appreciate your continued efforts to improve our manuscript. We have addressed your concerns as follows:

1. **Word Count Reduction:** We have significantly revised the manuscript to reduce redundancy, tighten language, and remove less crucial details. We are confident that the revised manuscript meets the JOSS word limit requirements.

2. **"State of the Field" Revision:** This section has been retitled and references have been updated to include more non-self citations, providing a broader perspective on the field.

3. **Figure Revisions:** We have increased the font size on the figure to ensure legibility. The Python and BioPython logos have been removed to avoid any implication of direct affiliation.

4. **Taxonomic Unit Clarification:** The term "genetic trees" refers to phylogenetic trees constructed from genetic data, where the taxonomic units are typically populations or species, not individual genes. This has been clarified in the text.

5. **"Climate Trees" Justification:** Constructing climate trees and performing tree distance comparisons provides a visual and quantitative way to assess the overall congruence between genetic relationships and climate similarities, complementing direct correlation tests and offering additional insights.

6. **Confounding Spatial Effects:** We acknowledge the potential confounding effects of spatial proximity on both genetic relatedness and climate similarity. Currently, our analyses do not explicitly control for spatial effects. We have added a discussion of this limitation and potential future directions for addressing it.

7. **Dependency Clarification:** The dependencies listed in the manuscript but not in setup.py are optional and not strictly required for end-users. We have clarified this in the text and removed specific version numbers.

8. **Minor Comments:** We have addressed all the minor comments, clarifying vague language, revising verb tense, and adding links to relevant GitHub content at specific commits.

We added iphylogeo++ ref into our manuscript, as suggested.

We believe that these revisions have significantly strengthened the manuscript and addressed all of your concerns. We are grateful for your comments and look forward to your further feedback.
