### 2024-06-11 #6579 @hazem-dev
We are pleased to inform you that we have thoroughly addressed your valuable feedback and implemented the suggested revisions in our manuscript. The revised version is now available on our GitHub repository.

In addition to addressing your [issue](https://github.com/tahiri-lab/aPhyloGeo/issues/49#issuecomment-2161600168), we have made significant efforts to streamline the manuscript, reducing its length by more than half and ensuring it now adheres to the JOSS word count limit. We believe these revisions have significantly strengthened the paper and enhanced its clarity.

Thank you, @mmore500 and @fboehm, again for your time and consideration.
