+++ aphylogeo/requirements.txt
@@ -1,9 +1,1 @@
-bio==1.5.2
-multiprocess==0.70.14
-pandas==1.3.5
-psutil==5.9.4
-PyYAML==6.0
-pytest==7.2.1
-numpy>=1.21.6
-cython==3.0.1
-git+https://github.com/althonos/pymuscle5
+aphylogeo==1.0.0
