+++ paper/paper.bib
@@ -215,7 +215,8 @@
   volume={7},
   pages={1--8},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+ doi = {10.1186/1471-2148-7-214}
 }
 @article{ronquist2012mrbayes,
   title={MrBayes 3.2: efficient Bayesian phylogenetic inference and model choice across a large model space},
@@ -225,7 +226,8 @@
   number={3},
   pages={539--542},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/sysbio/sys029}
 }
 @incollection{scott2009spatial,
   title={Spatial statistics in ArcGIS},
@@ -233,5 +235,6 @@
   booktitle={Handbook of applied spatial analysis: Software tools, methods and applications},
   pages={27--41},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-642-03647-7_2}
 }
