### 2024-07-01 #6579 @annazhukova
Thank you for the changes. However, I still do not understand from the article what exactly aPhylogeo does and why cannot this be achieved with other methods (e.g. with the [generalized liner model](https://journals.plos.org/plospathogens/article?id=10.1371/journal.ppat.1003932) in BEAST, which is not mentioned in the article).

@fboehm and the authors, I am sorry, but I will not have any time to review this further in the near future, so it would be the best if you find another reviewer.
