### 2024-07-01 #6579 @TahiriNadia
> Thank you for the changes. However, I still do not understand from the article what exactly aPhylogeo does and why cannot this be achieved with other methods (e.g. with the [generalized liner model](https://journals.plos.org/plospathogens/article?id=10.1371/journal.ppat.1003932) in BEAST, which is not mentioned in the article).
>
> @fboehm and the authors, I am sorry, but I will not have any time to review this further in the near future, so it would be the best if you find another reviewer.

Thank you, @annazhukova, for this note. However, the primary objectives of BEAST and aPhyloGeo are fundamentally different. BEAST is a cross-platform software designed for Bayesian analysis of molecular sequences using Markov Chain Monte Carlo (MCMC) methods, focusing exclusively on sequence data. In contrast, aPhyloGeo aims to elucidate the relationship between genetic evolution and environmental factors of species' habitats (e.g., temperature, humidity, soil composition, sunlight, precipitation, and others). The goal is to identify genetic signatures that explain species adaptation and biodiversity.
