### 2024-07-03 #6579 @fboehm
> Dear @fboehm,
>
> As it has been nearly three months (i.e. April, 3rd) since we submitted our article, I'm wondering if it is possible to resume work on the package without interfering with the review process. We paused development on the package during the last three months to avoid disrupting the reviewers. Thank you!
>
> Best.

@TahiriNadia - would it be possible for your development to proceed on a new git branch? That way, we could continue to point reviewers to the "main" branch; any issues that reviewers raise could be addressed with git pushes to the main branch. your other development activities could continue on another branch and would not be considered in this joss submission. Is that reasonable?

Following joss publication, you could, of course, merge the new development branch into main, so that users could get the latest updates.
