### 2024-07-03 #6579 @fboehm
> Thank you for the changes. However, I still do not understand from the article what exactly aPhylogeo does and why cannot this be achieved with other methods (e.g. with the [generalized liner model](https://journals.plos.org/plospathogens/article?id=10.1371/journal.ppat.1003932) in BEAST, which is not mentioned in the article).
>
> @fboehm and the authors, I am sorry, but I will not have any time to review this further in the near future, so it would be the best if you find another reviewer.

Thank you for the work in reviewing this submission, @annazhukova - i hope that we can work together on a future submission.
