### 2024-07-04 #6579 @TahiriNadia
> would it be possible for your development to proceed on a new git branch? That way, we could continue to point reviewers to the "main" branch; any issues that reviewers raise could be addressed with git pushes to the main branch. your other development activities could continue on another branch and would not be considered in this joss submission. Is that reasonable?
>
> Following joss publication, you could, of course, merge the new development branch into main, so that users could get the latest updates.

@fboehm - absolutely, that sounds like a great plan! I'll proceed with the development on a new git branch. Thank you!
