### 2024-07-04 27bd9f4
tutorials rephrase (readme)
