+++ paper/paper.md
@@ -52,13 +52,13 @@ Available as a PyPI package, *aPhyloGeo* enhances understanding of evolutionary 
 
 # Statement of Need
 
-Phylogeography, the study of the interplay between evolution and environment, has long been hindered by the challenges of integrating complex phylogenetic and environmental datasets. aPhyloGeo revolutionizes this field by offering a user-friendly Python package that streamlines phylogeographic analysis. 
+Phylogeography, the study of the interplay between evolution and environment, has long been hindered by the challenges of integrating complex phylogenetic and environmental datasets. aPhyloGeo contributes to this field by offering a user-friendly Python package that streamlines phylogeographic analysis. 
 By facilitating the investigation of the intricate relationship between genetic variation and environmental factors, aPhyloGeo is poised to advance our understanding of evolution, biogeography, and conservation efforts. This powerful package makes phylogeographic analysis more accessible and efficient, opening new avenues for research and discovery.
 The aPhyloGeo package is designed for versatile integration, offering flexibility to be incorporated into various environments. It can be seamlessly embedded within Python scripts for custom workflows, deployed on cloud platforms for scalable analysis, or integrated into web and desktop applications for user-friendly access. This adaptability makes it a valuable tool for diverse research and development needs, from individual projects to large-scale applications.
 
-# Statement of Field
+# Related Software Packages
 
-While BEAST [@drummond2007beast] and MrBayes [@ronquist2012mrbayes] are primarily used for phylogenetic analyses, they lack integrated tools for analyzing climatic data. Other software like ArcGIS [@scott2009spatial] focuses on geospatial analysis but may not be tailored for phylogeographic research. aPhyloGeo bridges this gap by providing a unified platform for both analyses, with a streamlined workflow.
+While BEAST [@drummond2007beast] and MrBayes [@ronquist2012mrbayes] are popular for phylogenetic analyses, they lack integrated tools for analyzing climatic data. Other software like ArcGIS [@scott2009spatial] focuses on geospatial analysis but may not be tailored for phylogeographic research. aPhyloGeo bridges this gap by providing a unified platform for both types of analysis, with a streamlined workflow.
 
 ## aPhyloGeo: A Software Solution
 
@@ -79,20 +79,18 @@ Given the urgency of climate change, new bioinformatics tools are needed. *aPhyl
 
 # Pipeline
 
-Navigating the *aPhyloGeo* workflow (refer to \autoref{fig:figure1}) is indispensable to fully harness the potential of this bioinformatics pipeline. The visual representation in \autoref{fig:figure1} outlines the key steps for conducting phylogeographic analysis with optimized effectiveness.
-For a more detailed understanding, please refer to the comprehensive tutorial [provided with the software](https://github.com/tahiri-lab/aPhyloGeo/wiki/Tutorial).
+\autoref{fig:figure1} outlines the key steps for conducting phylogeographic analysis using *aPhyloGeo*.
+For a more detailed understanding, please refer to the [comprehensive tutorial](https://github.com/tahiri-lab/aPhyloGeo/wiki/Tutorial) and [provided with the software](https://github.com/tahiri-lab/aPhyloGeo/wiki/Worflow).
 
 ![The workflow of the algorithm. The operations within this workflow include several blocks.\label{fig:figure1}](../img/workflow_en.png)
 
-The diagram below illustrates the workflow of the algorithm, consisting of several key blocks, each highlighted with a distinct color (refer to the [wiki page](https://github.com/tahiri-lab/aPhyloGeo/wiki/Worflow)).
-
 ## Multiprocessing
 
[…]
