### 2024-07-06 tahiri-lab/aPhyloGeo#49 @mmore500
Thank you for your comments.

I had the opportunity to look through the revised manuscript, and some significant improvements have been made. I opened #52 with some minor grammar and vocabulary suggestions.

To continue with some specific points.

> Figure Revisions: We have increased the font size on the figure to ensure legibility. The Python and BioPython logos have been removed to avoid any implication of direct affiliation.

I can see that the external logos have been removed.
However, I remain concerned with the font sizes, which do not seem to have significantly changed.
Here's a screen grab of a previous draft beside the current draft.

![Screenshot from 2024-07-05 21-33-34](https://github.com/tahiri-lab/aPhyloGeo/assets/10763333/7c9cbb66-7307-4f44-b202-78d527721ae3)

Ideally, the text in the figure should (at smallest) not be appreciably smaller than the figure caption font size.
The fact that *all* font sizes (e.g., including major titles) are much smaller than the caption font significantly contributes to the legibility issue here.

> Confounding Spatial Effects: We acknowledge the potential confounding effects of spatial proximity on both genetic relatedness and climate similarity. Currently, our analyses do not explicitly control for spatial effects. We have added a discussion of this limitation and potential future directions for addressing it.

I did not find this in the manuscript.

Quoting from the manuscript,
> The methods validate input parameters to ensure data accuracy and prevent errors (refer to the YAML file).

It is unclear what would be found in "the YAML file" without clicking through to see that it is a listing of parameters.

> Emphasizing software development best practices and open-source principles (e.g., iPhyloGeo++), aPhyloGeo ensures reliability and sets the stage for ongoing innovation.

Based on our discussion, I understand that the reference to iPhyloGeo++ is meant to demonstrate a "Project Using the Software" to evidence impact/adoption.
In the actual manuscript, it is unclear why iPhyloGeo++ is mentioned.

As an additional comment, this sentence has a lot of buzzwords but does not communicate a point that is specific and tangible.
This pattern arises (to varying degrees) throughout the manuscript.
This has improved somewhat in recent drafts, but I would encourage a careful pass through the text to consider this issue.

> Sequences with notable variability were specifically retained for analysis.

The tense of this sentence makes it sound like you are describing a specific project where an analysis was applied, instead of a general tool for analysis.

> aPhyloGeo employs algorithms using metrics like least squares (Felsenstein, 1997), Euclidean, and Robinson-Foulds (Robinson & Foulds, 1981) distances to ensure statistically sound correlations.

Why are these details relevant to mention in the abstract?
Consider clarifying or deleting.

>  In the comparison of phylogenetic trees, which are constructed based on genetic data, with climatic trees, a crucial step involves applying a phylogeography approach.

This is the only place that climatic trees are mentioned in the manuscript.
Based on previous drafts of the manuscript and our discussion, I understand this concept to be central to the aPhyloGeo workflow.
I am under the impression that the "climate tree" is novel methodology that your group has been involved in developing, and so is likely unfamiliar to much of the potential readership.
This method should be briefly defined, explained, and justified.
A reference to a publication introducing or applying the method would be highly beneficial.

## Additional comments

Please double-check your links by clicking through all of them in the PDF.
Some appear to be dead in the current draft.

There appear to be distinct comparison methods for trees and for sequences.
This could be clarified by specifying more explicitly what is being compared in the "Similarity Methods" section.
