### 2024-07-06 #6579 @mmore500
I have added some suggestions and further comments on the manuscript [here](https://github.com/tahiri-lab/aPhyloGeo/issues/49#issuecomment-2211587316).
