### 2024-07-06 #6579 @fboehm
@jdalapicolla @AgustinPardo @fredericlemoine - would you be willing to review this submission for JOSS? We are looking for a second reviewer.

Thank you for considering this!
