### 2024-07-08 #6579 @fboehm
@theosanderson @ammaraziz @hseabolt @richelbilderbeek @Gab0 - would you be willing to review this submission for JOSS? One of our original reviewers needed to discontinue, so we are in need of another reviewer.

Please feel free to reply in this comment thread below. Thank you for considering this!
