### 2024-07-08 #6579 @theosanderson
I am happy to review. (For info, I have not reviewed for JOSS before, so am slightly new to the process.)
