### 2024-07-08 b6e44e2
Update paper.md
add description to the YAML
