### 2024-07-09 #6579 @TahiriNadia
> @fboehm I am happy to review. (For info, I have not reviewed for JOSS before, so am slightly new to the process.)

Thanks @theosanderson.
