### 2024-07-09 tahiri-lab/aPhyloGeo#49 @TahiriNadia
> I had the opportunity to look through the revised manuscript, and some significant improvements have been made. I opened https://github.com/tahiri-lab/aPhyloGeo/pull/52 with some minor grammar and vocabulary suggestions.

Done :white_check_mark:
