### 2024-07-09 tahiri-lab/aPhyloGeo#49 @TahiriNadia
> I can see that the external logos have been removed.
> However, I remain concerned with the font sizes, which do not seem to have significantly changed.
> Here's a screen grab of a previous draft beside the current draft.

We have removed the logos and slightly increased the text size to prevent the image from becoming low quality and to ensure it fully contains all information. Additionally, we've made the figure accessible by including the image source (drawio) on GitHub, making it easier for the reader to understand. For enhanced readability, we recommend zooming in. We have also included a link to the image in the article. Additionally, we have provided a vectorial image in the Git repository. Thank you for your attention.
