### 2024-07-09 fee8e59
update Figure 1 JOSS
