### 2024-07-09 #6579 @fboehm
> @fboehm I am happy to review. (For info, I have not reviewed for JOSS before, so am slightly new to the process.)

Thank you, @theosanderson ! I'm excited for your first JOSS review. Our reviews are driven by checklists, so the first thing for you to do is to generate your personal checklist. It will guide your review. To generate the checklist, you need to write a comment in this comment thread. It should contain only the text:

Once you've posted this comment, you'll see that the "editorialbot" will post a new comment that contains a series of boxes. You can then assess each item; if you feel that the submission meets the stated criteria, put a check in the corresponding box; otherwise, write a comment in this thread (or open a github "issue" in the submitted repository) to describe your concern and suggest changes that the author can make so that the result will meet the criteria.

You can always ask me (@fboehm) if you're unsure about anything or if you feel that a stated criterion needs clarification.
