### 2024-07-09 tahiri-lab/aPhyloGeo#49 @TahiriNadia
```
To continue with some specific points,

Figure Revisions: We have increased the font size on the figure to ensure legibility. The Python and BioPython logos have been removed to avoid any implication of direct affiliation.

I can see that the external logos have been removed.
However, I remain concerned with the font sizes, which do not seem to have significantly changed.
Here's a screen grab of a previous draft beside the current draft.

Screenshot from 2024-07-05 21-33-34
![image](https://github.com/tahiri-lab/aPhyloGeo/assets/19578926/6869a106-d975-4ecf-b1eb-607fc5b0b023)

Ideally, the text in the figure should (at smallest) not be appreciably smaller than the figure caption font size.
The fact that all font sizes (e.g., including major titles) are much smaller than the caption font significantly contributes to the legibility issue here.
```

Done ✅. We completely updated the new figure (see below) with particular attention to the text size. It is now more accessible. Thanks.

![image](https://github.com/tahiri-lab/aPhyloGeo/assets/19578926/04987d74-1cea-4da6-ad29-c33015ff635b)
