### 2024-07-09 aebab0c
add spatial proximity confounding Effects
