### 2024-07-09 tahiri-lab/aPhyloGeo#49 @hazem-dev
@mmore500 ,

Thank you for your thorough and insightful review of our manuscript. We greatly appreciate the time and effort you have invested in providing valuable feedback.

We have carefully addressed each of your comments in our revised manuscript. Below, we provide a detailed response to each of your points and describe the corresponding changes made:

1. **Confounding Spatial Effects**:
   We have now included a section discussing the potential confounding effects of spatial proximity on both genetic relatedness and climate similarity. This limitation and possible future directions for addressing it, such as incorporating spatial autocorrelation analysis and spatial regression models, are now explicitly mentioned under the "Spatial Proximity Confounding Effects" section.

2. **YAML File Reference**:
   We have clarified the contents and purpose of the YAML configuration file. The methods section now details the types of parameters listed in the YAML file, providing a clearer understanding without the need to click through the link.

3. **Reference to iPhyloGeo++**:
   The reference to iPhyloGeo++ has been rephrased to clearly illustrate its significance as a "Project Using the Software" to evidence impact and adoption. This section now underscores the practical application and impact of aPhyloGeo in phylogeographic research, enhancing its context and relevance.

4. **Buzzwords and Specificity**:
   We have conducted a careful review of the manuscript to eliminate overly broad statements and buzzwords.

5. **Sentence Tense**:
   The sentence "Sequences with notable variability were specifically retained for analysis" has been revised to "Sequences with notable variability are specifically retained for analysis" to reflect the general applicability of the tool rather than a specific project.

6. **Abstract Details**:
   The details regarding the algorithms employed by aPhyloGeo have been streamlined in the abstract to focus on the core functionality and benefits of the software.

7. **Climatic Trees Explanation**:
   We have introduced a brief definition and explanation of climatic trees, a novel methodology central to the aPhyloGeo workflow. This includes a description of how climatic trees are constructed and their significance in phylogeographic analysis.

8. **Link Verification**:
   We have thoroughly checked all links in the manuscript to ensure they are functional and lead to the correct resources.

9. **Clarification of Comparison Methods**:
   The distinction between tree and sequence comparison methods has been clarified in the "Similarity Methods" section.

Thank you once again for your valuable comments and for considering our revised manuscript. We look forward to your favorable response.
