### 2024-07-09 #6579 @hazem-dev
For [@mmore500](https://github.com/mmore500) and [@fboehm](https://github.com/fboehm) - We are pleased to confirm that all [issues](https://github.com/tahiri-lab/aPhyloGeo/issues/49) raised during the review process have been fully addressed.
