### 2024-07-09 #6579 @fboehm
> For [@mmore500](https://github.com/mmore500) and [@fboehm](https://github.com/fboehm) - We are pleased to confirm that all [issues](https://github.com/tahiri-lab/aPhyloGeo/issues/49) raised during the review process have been fully addressed.

Thank you, @hazem-dev and @TahiriNadia . @mmore500 @theosanderson - please feel free to examine the revisions as you continue the reviews. For @theosanderson - this will just be the "current" state of the main git branch of the repository. @mmore500 - please feel free to consider the authors' efforts to address your suggestions and to clarify remaining concerns. Thank you again!
