### 2024-07-09 #6579 @theosanderson
I have not yet assessed every item on the checklist, but I echo the other reviewers on the point that the manuscript at present lacks a clear example to demonstrate the usefulness of this tool: https://github.com/tahiri-lab/aPhyloGeo/issues/55
