### 2024-07-10 #6579 @TahiriNadia
> I have not yet assessed every item on the checklist, but I echo the other reviewers on the point that the manuscript at present lacks a clear example to demonstrate the usefulness of this tool: [tahiri-lab/aPhyloGeo#55](https://github.com/tahiri-lab/aPhyloGeo/issues/55)

Thanks @theosanderson. We'll be working on it, especially the new data set you suggested incorporating as a validation step for our software.
