### 2024-07-21 #6579 @mmore500
Hi all— just wanted to post a quick update that I’m wrapping up the last of my international travel for the summer this week and have time scheduled on that time frame to return and update my review. Looking forward to it, thanks!
