### 2024-08-06 #6579 @mmore500
Looking through the repository, I wanted to surface some brief concerns about licensing.

I see that aPhyloGeo is licensed under the MIT license. However, within the `aphylogeo/bin` folder there are binaries for muscle, and fastree --- which appear to be GPL licensed. Clustalw2 is LGPL licensed. My understanding is that due to its ``copyleft'' nature, the GPL imposes certain restrictions requiring that derivative works and redistributions of GPL-licensed works must also themselves be GPL-licensed.

Additionally, there are binaries for MAFFT, which is BSD licensed. My understanding is that BSD-licensed code must contain certain license notices, e.g.,
```
Redistributions of source code must retain themay at the very least
above copyright notice, this list of conditions
and the following disclaimer.  Redistributions in
binary form must reproduce the above copyright
notice, this list of conditions and the following
disclaimer in the documentation and/or other
materials provided with the distribution.
```
I believe that it may be necessary to include license notices for all binaries bundled with aPhyloGeo.

In the case that aPhyloGeo remains MIT licensed, it may be necessary to have users supply their own binaries for GPL-licensed programs --- but I am unsure on this point. Alternatively, it might be possible to resolve these concerns by licensing aPhyloGeo as GPL and including license notices for bundled binaries.

I am not a lawyer, though, and would not be comfortable making a final determination on any of the above without outside input. :+1:
