### 2024-08-12 #6579 @fboehm
@theosanderson - do you haave suggestions for handling the licensing issue?
