### 2024-08-14 f0d2d56
Changement de licenses pour expliciter celles externes + modifications des unit tests pour etre agnostique a l'emplacement du navigateur de fichiers
