+++ pyproject.toml
@@ -33,6 +33,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
 pythonpath = "aphylogeo/"
+testpaths = 'tests/'
 
 [tool.ruff]
 line-length = 130
