### 2024-08-21 #6579 @theosanderson
Sorry for the delay and for dropping off this thread. I have now completed my checklist. My primary concern continues to be that I don't feel that I've seen evidence that this tool is able to discover scientifically meaningful results. (I'm not asserting that it cannot, just that I haven't seen evidence). I set out these concerns in https://github.com/tahiri-lab/aPhyloGeo/issues/55 .
