### 2024-09-10 #6579 @TahiriNadia
> Sorry for the delay and for dropping off this thread. I have now completed my checklist. My primary concern continues to be that I don't feel that I've seen evidence that this tool is able to discover scientifically meaningful results. (I'm not asserting that it cannot, just that I haven't seen evidence). I set out these concerns in [tahiri-lab/aPhyloGeo#55](https://github.com/tahiri-lab/aPhyloGeo/issues/55) .

No problem at all, and thanks for following up! Instead of using the dataset you originally suggested (due to some links being unavailable), we shifted our focus to the dataset from Uhlir et al. (2021) "Adding pieces to the puzzle: insights into diversity and distribution patterns of Cumacea (Crustacea: Peracarida) from the deep North Atlantic to the Arctic Ocean" [1]. Through this analysis, we derived substantial conclusions that demonstrate the scientific relevance and potential of the package. These findings offer compelling evidence of its utility for the research community.

Additionally, we are currently in active communication with the lead author of the Uhlir et al. study to ensure the final validation of our results. You can find further details and examples of our study case in the [documentation](https://github.com/tahiri-lab/aPhyloGeo/wiki/Study-case).

I hope this provides clarity, and I'm happy to engage in further discussion if needed!

[1] Uhlir, C., Schwentner, M., Meland, K., Kongsrud, J. A., Glenner, H., Brandt, A., Thiel, R., Svavarsson, J., Lörz, A.-N., & Brix, S. (2021). Adding pieces to the puzzle: insights into diversity and distribution patterns of Cumacea (Crustacea: Peracarida) from the deep North Atlantic to the Arctic Ocean. PeerJ, 9, e12379.
