### 2024-09-10 #6579 @TahiriNadia
> @TahiriNadia - do you have any questions about the remaining tasks?
>
> Please let me know. Thanks again!

> @TahiriNadia - do you have any questions about how to address the remaining open issues?

Thank you, @fboehm. I’ve completed the revisions for the second tutorial as recommended in the review and am now addressing licensing and making a few final adjustments. I’ll present a new solution ASAP.
