### 2024-09-20 d88224a
Run the application cmd line
