### 2024-09-24 d74660a
Added Sphinx documentation to the project
