+++ docs/source/api/multiProcessor.rst
@@ -1,0 +1,9 @@
+.. _aphylogeo.multiProcessor:
+
+aphylogeo.multiProcessor
+========================
+
+.. automodule:: aphylogeo.multiProcessor
+    :members:
+    :undoc-members:
+    :show-inheritance:
