+++ docs/source/installation.rst
@@ -1,0 +1,41 @@
+.. _installation:
+
+Installation
+============
+
+Linux UNIX & Windows versions
+-----------------------------
+
+`aPhyloGeo` is available as a Python package.
+
+Installation from PyPI via pip
+------------------------------
+
+`aPhyloGeo` can be installed directly from the Python Packaging Index:
+
+.. code-block:: bash
+
+    python3 -m pip install aPhyloGeo
+
+Installation from Source via Poetry
+-----------------------------------
+
+This package uses `Poetry` dependency management and packaging tool for Python. 
+Poetry installation guide can be found here: `Poetry Install <https://python-poetry.org/docs/#installation>`_.
+
+.. warning::
+
+   For Windows installation, it is recommended to launch PowerShell in **Administrator mode**.
+
+Installation of Poetry is done using the following command:
+
+.. code-block:: bash
+
+    pip install poetry
+
+Once Poetry is installed, you can clone the repository and install the package using the following commands:
+
+.. code-block:: bash
+
+    poetry install
+
