+++ docs/modules.rst
@@ -1,0 +1,7 @@
+aphylogeo
+=========
+
+.. toctree::
+   :maxdepth: 4
+
+   aphylogeo
