+++ docs/source/sequence-similarity-algorithms.rst
@@ -1,0 +1,101 @@
+.. _sequence-similarity-algorithms:
+
+Sequence Similarity Algorithms
+==============================
+
+Filtering Alignment Windows
+---------------------------
+
+To enhance the algorithm's efficiency, an additional step has been introduced. This step filters alignment windows, retaining only those with a dissimilarity rate surpassing the user-defined threshold specified in the parameter YAML file. These retained windows represent the most significant sequence alignments.
+
+Edit-Based Algorithms in Bioinformatics and Phylogeny
+-----------------------------------------------------
+
+Edit-based algorithms, often referred to as distance-based algorithms, quantify the minimum number of single-character operations (insertions, deletions, or substitutions) needed to convert one sequence into another. The greater the number of operations required, the lower the similarity or greater the distance between the sequences. In the field of bioinformatics, these algorithms play a crucial role in tasks related to phylogeny and phylogeography, enabling the comparison of genetic sequences and their evolutionary relationships. They serve as a fundamental metric for various sequence similarity techniques and find widespread application in bioinformatics tasks such as DNA sequence alignment, phylogenetic analysis, and molecular evolution studies.
+
+Hamming Distance
+~~~~~~~~~~~~~~~~
+
+The Hamming distance, denoted as :math:`H(s_1, s_2)`, measures the number of positions at which the corresponding symbols differ between two equal-length sequences. It is suitable for sequences of equal length and does not consider insertions or deletions.
+
+.. math::
+
+    H(s_1, s_2) = \sum_{i=1}^{n} \delta(s_{1i} \neq s_{2i})
+
+Levenshtein Distance
+~~~~~~~~~~~~~~~~~~~~
+
+The Levenshtein distance, denoted as :math:`L(s_1, s_2)`, calculates the minimum number of single-character edits (insertions, deletions, or substitutions) needed to transform one sequence into another. It is applicable to sequences of varying lengths.
+
+.. math::
+
+    L(s_1, s_2) = \min\left( L(s_1[1:m-1], s_2) + 1, L(s_1, s_2[1:n-1]) + 1, L(s_1[1:m-1], s_2[1:n-1]) + \delta(s_{1m} \neq s_{2n}) \right)
+
+Damerau-Levenshtein Distance
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The Damerau-Levenshtein distance, denoted as :math:`D(s_1, s_2)`, extends the Levenshtein distance by including the transposition of adjacent characters as a valid operation. It is useful for detecting both insertions and transpositions.
+
+.. math::
+
+    D(s_1, s_2) = \min\left( D(s_1[1:m-1], s_2) + 1, D(s_1, s_2[1:n-1]) + 1, D(s_1[1:m-1], s_2[1:n-1]) + \delta(s_{1m} \neq s_{2n}), D(s_1[1:m-2], s_2[1:n-2]) + 1 \right)
+
+Jaro Similarity
+~~~~~~~~~~~~~~~
+
+Jaro similarity, denoted as :math:`J(s_1, s_2)`, measures the similarity between two strings by considering the number of matching characters and the number of transpositions. It is suitable for comparing strings with small differences and is particularly effective for short strings.
+
+.. math::
+
+    J(s_1, s_2) = \frac{1}{3} \left( \frac{m}{\lvert s_1 \rvert} + \frac{m}{\lvert s_2 \rvert} + \frac{m - t}{m} \right)
+
+.. math::
+
[…]

[…]
