### 2024-09-24 a213d14
Formatted functions documentation for the Sphinx autodoc to be cleaner
