+++ aphylogeo/utils.py
@@ -36,17 +36,15 @@ def header():
 
 def getDissimilaritiesMatrix(df, columnWithSpecimenName, columnToSearch):
     """
-    Creation of a list containing the names of specimens and minimums
-    tempratures
+    Create a list containing the names of specimens and their corresponding minimum temperatures.
 
-    Args:
-        df (content of CSV file)
-        columnWithSpecimenName (first column of names)
-        columnToSearch (column to compare with the first one)
+    :param df: The content of the CSV file.
+    :param columnWithSpecimenName: The first column containing specimen names.
+    :type columnWithSpecimenName: str
+    :param columnToSearch: The column to compare with the first one.
+    :type columnToSearch: str
 
-    Return:
-        The dissimilarities matrix
-
+    :return: The dissimilarities matrix.
     """
     meteoData = df[columnToSearch].tolist()
     nomVar = df[columnWithSpecimenName].tolist()
@@ -86,21 +84,24 @@ def getDissimilaritiesMatrix(df, columnWithSpecimenName, columnToSearch):
 
 def leastSquare(tree1, tree2):
     """
-    Method that calculates the least square distance between two trees.
-    Trees must have the same number of leaves.
-    Leaves must all have a twin in each tree.
-    A tree must not have duplicate leaves
-     x   x
-    ╓╫╖ ╓╫╖
-    123 312
+    Calculate the least square distance between two trees.
 
-    Args:
-        tree1 (distanceTree object from biopython)
-        tree2 (distanceTree object from biopython)
+    - Trees must have the same number of leaves. 
+    - Leaves must all have a twin in each tree. 
+    - A tree must not have duplicate leaves.
 
-    Return:
-        return result (double) the final distance between the two
+    Example:
+        x      x \n  
+       ╓╫╖    ╓╫╖ \n
+       123    312 
 
+    :param tree1: The first tree to compare.
+    :type tree1: distanceTree (from Biopython)
+    :param tree2: The second tree to compare.
+    :type tree2: distanceTree (from Biopython)
+
+    :return: The final distance between the two trees.
+    :rtype: float
     """
     ls = 0.00
     leaves = tree1.get_terminals()
@@ -118,20 +119,23 @@ def leastSquare(tree1, tree2):
 def robinsonFoulds(tree1, tree2):
     """
     Method that calculates the robinson foulds distance between two trees.
-    Trees must have the same number of leaves.
-    Leaves must all have a twin in each tree.
-    A tree must not have duplicate leaves
-     x   x
-    ╓╫╖ ╓╫╖
+    
+    - Trees must have the same number of leaves.
+    - Leaves must all have a twin in each tree.
+    - A tree must not have duplicate leaves
+    
+    Example:
+     x   x \n
+    ╓╫╖ ╓╫╖ \n
     123 312
 
-    Args:
-        tree1 (distanceTree object from biopython converted to Newick)
-        tree2 (distanceTree object from biopython converted to Newick)
-
-    Return:
-        return result the final distance between the two
-
+    :param tree1: The first tree to compare
+    :type tree1: distanceTree object from biopython converted to Newick
+    :param tree2: The second tree to compare
+    :type tree2: distanceTree object from biopython converted to Newick
+    
+    :return: The final distance between the two
+    :rtype: float
     """
     rf = 0
     tree1_newick = ete3.Tree(tree1.format("newick"), format=1)
@@ -153,17 +157,17 @@ def euclideanDist(tree1, tree2):
     The bipartition bitmasks of the trees must be correct for the current tree
     structures (by calling :meth:`Tree.encode_bipartitions()` method)
 
-     x   x
-    ╓╫╖ ╓╫╖
[…]
