### 2024-09-25 tahiri-lab/aPhyloGeo#55 @theosanderson
Thank you @TahiriNadia,

Could you explain in simpler words what aPhyloGeo has shown here? One thing I haven't got my head around is what the Euclidean distances referred to are - i.e. the euclidean distance between what and what?

More broadly, it just seems very unlikely to me that 400 bp sequences of ribosomal RNA would have genuine causative interactions with wind speed or oxygen concentration?
