### 2024-10-04 aa35656
Cleaned .rst mathematical equations format
