### 2024-10-24 0b742a7
Removed the version name of the muscle executables and added the documentation workflow
