+++ .github/workflows/documentation.yml
@@ -1,0 +1,43 @@
+name: "aPhyloGeo CI/CD Pipeline"
+
+on:
+    push:
+        branches:
+            - main
+            - develop
+    pull_request:
+        branches:
+            - main
+            - develop
+    release:
+        types: [ created ]
+
+permissions:
+    contents: write
+
+jobs:
+    build-deploy-docs:
+        needs: security-scan
+        runs-on: ubuntu-latest
+        steps:
+            -   uses: actions/checkout@v4
+            -   name: Set up Python
+                uses: actions/setup-python@v4
+                with:
+                    python-version: '3.10'
+            -   name: Install dependencies
+                uses: py-actions/py-dependency-install@v4
+                with:
+                    path: "aphylogeo/requirements.txt"
+            -   name: Sphinx build
+                run: |
+                    sphinx-build docs _build
+            -   name: Deploy to GitHub Pages
+                uses: peaceiris/actions-gh-pages@v3
+                if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
+                with:
+                    publish_branch: gh-pages
+                    github_token: ${{ secrets.GITHUB_TOKEN }}
+                    publish_dir: _build/
+                    force_orphan: true
+
