### 2024-10-24 9aee867
Remove a dependency from the documentation workflow
