### 2024-10-24 1f61c2c
Added sphinx dependecies to requirements-dev file
