### 2024-10-24 c751460
Removed requirements-dev dependecies and added them directly to the workflow
