+++ .github/workflows/documentation.yml
@@ -24,13 +24,14 @@ jobs:
                 uses: actions/setup-python@v4
                 with:
                     python-version: '3.10'
-            -   name: Install dependencies
-                uses: py-actions/py-dependency-install@v4
-                with:
-                    path: "requirements-dev.txt"
+            -   name: Install Sphinx and sphinx-rtd-theme
+                run: |
+                    python -m pip install --upgrade pip
+                    pip install Sphinx sphinx-rtd-theme
             -   name: Sphinx build
                 run: |
-                    sphinx-build docs _build
+                    cd aphylogeo
+                    sphinx-build docs/source _build
             -   name: Deploy to GitHub Pages
                 uses: peaceiris/actions-gh-pages@v3
                 if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
