+++ .github/workflows/documentation.yml
@@ -30,7 +30,6 @@ jobs:
                     pip install Sphinx sphinx-rtd-theme
             -   name: Sphinx build
                 run: |
-                    cd aphylogeo
                     sphinx-build docs/source _build
             -   name: Deploy to GitHub Pages
                 uses: peaceiris/actions-gh-pages@v3
@@ -38,6 +37,6 @@ jobs:
                 with:
                     publish_branch: gh-pages
                     github_token: ${{ secrets.GITHUB_TOKEN }}
-                    publish_dir: aphylogeo/_build/
+                    publish_dir: _build/
                     force_orphan: true
 
