### 2024-10-24 634cc61
Fixed a problem where the api documentation on the github pages was missing
