+++ .github/workflows/documentation.yml
@@ -28,6 +28,7 @@ jobs:
                 run: |
                     python -m pip install --upgrade pip
                     pip install Sphinx sphinx-rtd-theme
+                    pip install -r aphylogeo/requirements.txt 
             -   name: Sphinx build
                 run: |
                     sphinx-build docs/source _build
