### 2024-10-24 3df9a3d
Removed muscle version from variables
