### 2024-10-25 e2ea9ac
Removed executables from the bin folder, added a guide to manually install the executables in the github page
