+++ .github/workflows/documentation.yml
@@ -26,9 +26,11 @@ jobs:
                     python-version: '3.10'
             -   name: Install Sphinx and sphinx-rtd-theme
                 run: |
-                    python -m pip install --upgrade pip
-                    pip install Sphinx sphinx-rtd-theme
-                    pip install -r aphylogeo/requirements.txt 
+                    python -m pip install --upgrade pip && \
+                    pip install -e . && \
+                    pip install Sphinx sphinx-rtd-theme && \
+                    pip install -r aphylogeo/requirements.txt
+                    
             -   name: Sphinx build
                 run: |
                     sphinx-build docs/source _build
