### 2024-10-25 8c3e6dd
Commented tests that depended on the executables inside the bin folder that are now removed
