### 2024-10-25 095ea2d
Formatted some of the documentation to be cleaner
