+++ aphylogeo/alignement.py
@@ -315,11 +315,10 @@ class AlignSequences:
     def muscleAlign(self):
         """Method to perform a multiple DNA sequence alignment using Muscle Algorithm
 
-        Return:
-        -------
-        (Dict): heuristicMSA
-            Keys: accession ID
-            Values: Aligned sequences
+        Returns (Dict):
+            heuristicMSA
+                - Keys: accession ID
+                - Values: Aligned sequences
         """
         try:
             if sys.platform == "win32":
@@ -343,11 +342,10 @@ class AlignSequences:
     def clustalAlign(self):
         """Method to perform a multiple DNA sequence alignment using ClustalW2 Algorithm
 
-        Return:
-        -------
-        (Dict): heuristicMSA
-            Keys: accession ID
-            Values: Aligned sequences
+        Returns (Dict):
+            heuristicMSA
+                - Keys: accession ID
+                - Values: Aligned sequences
         """
         try:
             if sys.platform == "win32":
@@ -369,11 +367,10 @@ class AlignSequences:
     def mafftAlign(self):
         """Method to perform a multiple DNA sequence alignment using MAFFT Algorithm
 
-        Return:
-        -------
-        (Dict): heuristicMSA
-            Keys: accession ID
-            Values: Aligned sequences
+        Returns (Dict):
+            heuristicMSA
+                - Keys: accession ID
+                - Values: Aligned sequences
         """
         try:
             if sys.platform == "win32":
@@ -421,16 +418,14 @@ class AlignSequences:
         The length of each sequence from the pairwise alignment are set equal by
         inserting dash (-) in most appropriate location of a given sequence.
 
-        Parameters:
-        -----------
-        alignment: dict of nested dict
-            {accession couple #1 : {Centroid Acc:Centroid Aligned Seq, Non-centroid Acc #1: non-centroid Aligned Seq #1},
-            ... ,
-            {accession couple #n : {Centroid Acc:Centroid Aligned Seq, Non-centroid Acc #n: non-centroid Aligned Seq #n}}
+        args:
+            alignment: dict of nested dict
+                {accession couple #1 : {Centroid Acc:Centroid Aligned Seq, Non-centroid Acc #1: non-centroid Aligned Seq #1},
+                ... ,
+                {accession couple #n : {Centroid Acc:Centroid Aligned Seq, Non-centroid Acc #n: non-centroid Aligned Seq #n}}
 
-        Return:
-        -------
-        A dictionary of all accessions and their fitted aligned sequences.
+        Returns:
+            A dictionary of all accessions and their fitted aligned sequences.
         """
         seqs = self.getAlignSeqs(aligned)
         max_len = max(self.getAlignSeqLens(aligned))
@@ -444,13 +439,11 @@ class AlignSequences:
     def getAlignSeqs(self, aligned):
         """Extract all sequences aligned using a pairwise alignment
 
-        Parameters:
-        -----------
-        alignment: see fitPairwise(alignment) docstring
+        args:
+            alignment: see fitPairwise(alignment) docstring
 
-        Return:
-        -------
-        List of sequences aligned through pairwise alignment
+        Returns:
+            List of sequences aligned through pairwise alignment
         """
         seqs = []
         for alignment in aligned:
@@ -460,42 +453,36 @@ class AlignSequences:
     def getAlignSeqLens(self, aligned):
         """Get length of all sequences aligned using a pairwise alignment
 
-        Parameters:
-        -----------
-        alignment: see fitPairwise(alignment) docstring
+        args:
+            alignment: see fitPairwise(alignment) docstring
 
-        Return:
-        -------
-        List of the length of each aligned sequences
+        Returns:
+            List of the length of each aligned sequences
         """
[…]
