### 2024-11-01 #6579 @arfon
:wave: folks, @fboehm has asked me to help out with his submissions so I'm going to take this one over.

@TahiriNadia – from your most recent comment here it seems like you feel like you have accommodated/responded to all of the reviewer feedback at this point?

Reading the thread history, the most meaningful outstanding question seems to be that from @theosanderson here: https://github.com/tahiri-lab/aPhyloGeo/issues/55 . @theosanderson do you feel like your questions/concerns have been adequately addressed?
