### 2024-11-01 #6579 @theosanderson
Hello,
I am grateful to the authors for their considerable work to respond to the reviewers comments.

Unfortunately, my initial concerns persist. I was concerned that the approach taken by aPhyloGeo would not yield meaningful results about connections between climatic factors and evolution. I have not yet seen evidence that it can. The authors have provided an analysis of 35 sequences of 500bp of ribosomal RNA from shrimps, and claim that they can identify specific regions of this sequence which correlate with wind-speed and (separately) with oxygen concentration. It seems to me very unlikely that an analysis of 35 sequences would have adequate power to detect statistically meaningful such effects giving the number of degrees of freedom (many windows of this sequence, many environmental variables). The authors' other example, of SARS-CoV-2, consists of just five sequences. (These concerns should not really be interpreted as a request for larger analyses: the convincing evidence would come from comparing to data with a known ground-truth relationship between evolution and climatic factors).
