### 2024-11-01 #6579 @arfon
Thanks @theosanderson. If I'm understanding the concern here, the concern is whether this software can be used to generate meaningful scientific results?

Typically we wouldn't try and pass judgement this at JOSS (i.e., we're not looking to validate scientific results here), but if I understand correctly, this software has yet to be used to generate scientific outputs for any peer-reviewed papers @TahiriNadia? I note that in [this issue](https://github.com/tahiri-lab/aPhyloGeo/issues/55), you point to this analysis https://github.com/tahiri-lab/aPhyloGeo/wiki/Study-case – is it correct to say that this analysis has not been published in a peer reviewed journal yet?
