### 2024-11-01 #6579 @theosanderson
Yes @arfon I think that's largely correct. My concerns relate to the claims that the tool can give insights into the role of climate in affecting evolution rather than into whether it performs the workflow in Fig. 1. I can confirm that it does (appear to) perform the workflow shown in Fig. 1, and if that's sufficient for an editorial decision that's great.

I believe the authors have cited aPhyloGeo in some reviewed conference proceedings.
