### 2024-11-06 #6579 @arfon
Thanks @theosanderson. @TahiriNadia – I'd like to offer you the chance to provide a clear rebuttal here, but based on the review feedback, my current inclination is to reject this submission due to lack of (peer reviewed) evidence that the software is doing something with scientific merit. Please respond with any further information you have.
