### 2024-11-06 #6579 @TahiriNadia
Thank you, @arfon, for providing the opportunity to elaborate on our work.

Our tool has been rigorously peer-reviewed and published in the SciPy proceedings over three consecutive years (2022, 2023, and most recently a few weeks ago in 2024), demonstrating a commitment to reproducible and scalable phylogenetic and phylogeographic analyses. This progression has consistently showcased improvements and new features that enhance the tool’s ability to address key challenges in host-virus relationships and pathogen evolution, as demonstrated in our recent work on coronavirus and bat host interactions (Li & Tahiri, 2024). Each publication highlights the practical utility and scientific relevance of our tool, addressing issues of data integration and reproducibility that are critical to advancing research in phylogeny.

Moreover, our tool’s pertinence has been further validated through an active collaboration with Uhlir et al. (2021), specifically leveraging their extensive dataset on Cumacea diversity and distribution from the North Atlantic to the Arctic, published in PeerJ. This collaborative validation provided an ecological framework that underscored aPhyloGeo’s flexibility and robustness when applied to complex ecological data, establishing its value beyond viral phylogenies and into broader ecological contexts. Uhlir et al.’s involvement has been instrumental in ensuring the tool’s functionality aligns with both ecological and evolutionary research needs.

In addition to these peer-reviewed outcomes, we are preparing a new manuscript that will present additional applications and validation results in a forthcoming journal submission. This new work will continue to build on our tool’s scientific impact, further solidifying its role as a versatile, scientifically robust resource for the research community.
