### 2024-11-06 #6579 @TahiriNadia
> Thanks @theosanderson. If I'm understanding the concern here, the concern is whether this software can be used to generate meaningful scientific results?
>
> Typically we wouldn't try and pass judgement this at JOSS (i.e., we're not looking to validate scientific results here), but if I understand correctly, this software has yet to be used to generate scientific outputs for any peer-reviewed papers @TahiriNadia? I note that in [this issue](https://github.com/tahiri-lab/aPhyloGeo/issues/55), you point to this analysis https://github.com/tahiri-lab/aPhyloGeo/wiki/Study-case – is it correct to say that this analysis has not been published in a peer reviewed journal yet?

Thanks, @arfon. To clarify, the tool has indeed been validated and used to produce scientific results that have undergone peer review. Recently, our results were accepted in the SciPy proceedings as part of a peer-reviewed conference, where we detailed the scientific insights generated by the tool. Additionally, we have actively collaborated with the Uhlir team, using their ecological datasets to further validate the software’s relevance and applicability.
