### 2024-11-07 #6579 @TahiriNadia
> The references provided are very useful, thanks for passing them along.
>
> In (Li & Tahiri, 2024), I do see how distance comparisons between trees (i.e., the bat host phylogenies and coronavirus phylogenies) were applied shed light on co-evolution. Would you be able to clarify the role of biogeography, and more explicitly aPhyloGeo’s concept of climatic trees, in that work?
>
> Could you also clarify the role that the aPhyloGeo code itself in that work? I looked through the workflow files linked at https://github.com/tahiri-lab/aPhyloGeo.sm/, and didn’t see aPhyloGeo used as an import or portions of the aPhyloGeo library code I recognized included directly.

Dear @mmore500, Thank you for the insightful question. For our study, we did indeed include the _aPhyloGeo_ package via pipy to support the analyses.

The references you provided offer valuable context, particularly regarding the role of distance comparisons in understanding co-evolution patterns, as seen in (Li & Tahiri, 2024). In that work, _biogeography_, and more specifically the concept of _climatic trees_ within **aPhyloGeo**, was essential for mapping phylogenetic relationships against climatic gradients, revealing potential ecological constraints on co-evolutionary dynamics.

Best.
