### 2024-11-07 #6579 @mmore500
I do see some mentions of the bat populations being sampled from a specific region (China). I did also see a brief note in the discussion that suggests an analogy between vicariance/biogeography and host range:

> When contemplating the associations between hosts and parasites, particularly within the scope of our study of CoVs, an analogy can be drawn from vicariance biogeography. In this analogy, the host can be likened to a “region” exploited by the parasite, and host speciation can be seen as analogous to a “vicariance event” [58].

Could you more directly point me to which analyses reported in <https://doi.org/10.3390/v16071133> explicitly incorporate climactic or geographic data? I also looked in the supplement at <https://www.mdpi.com/article/10.3390/v16071133/s1> but did not immediately see where climactic or geographic data was incorporated into analyses.
