### 2024-11-07 #6579 @TahiriNadia
> Looking closely through (Li & Tahiri, 2024), I do see some mentions of the bat populations being sampled from a specific region (China). I did also see a brief note in the discussion that suggests an analogy between vicariance/biogeography and host range:
>
> > When contemplating the associations between hosts and parasites, particularly within the scope of our study of CoVs, an analogy can be drawn from vicariance biogeography. In this analogy, the host can be likened to a “region” exploited by the parasite, and host speciation can be seen as analogous to a “vicariance event” [58].
>
> Could you more directly point me to which analyses reported in https://doi.org/10.3390/v16071133 explicitly incorporate climactic or geographic data? I also looked in the supplement at https://www.mdpi.com/article/10.3390/v16071133/s1 but did not immediately see where climactic or geographic data was incorporated into analyses.

Thank you, @mmore500, for your detailed review and questions. In (Li & Tahiri, 2024), the primary focus was indeed on the cophylogenetic patterns between coronaviruses (CoVs) and bats rather than on explicit climatic or geographic variables. The geographic scope—bat samples from China—was contextualized primarily to leverage the region’s diverse bat populations as a rich source of CoV diversity and host-virus interactions. Specific geographic origins of the bat species were referenced through metadata in GenBank and the original studies; however, these were not analyzed with climate data or mapped with explicit spatial models (but indirectly with geographical conditions).

The study does make an analogy to vicariance biogeography in discussing host-parasite relationships, but this analogy serves more as a conceptual framework rather than being operationalized in a climate or geography-based model. The analyses center on genetic data, phylogenetic congruence, geographic, and recombination regions, focusing on the molecular dynamics between hosts and CoVs rather than incorporating climatic variables directly.

If further elaboration on these regional influences is valuable, incorporating spatial or climate-focused data could indeed provide additional context to the observed host-virus associations.
