### 2024-11-07 #6579 @TahiriNadia
The most recent peer-reviewed study directly addressing Cumacea is as follows:

4️⃣ Gagnon, J. & Tahiri, N. (2024). Ecological and Spatial Influences on the Genetics of Cumacea (Crustacea: Peracarida) in the Northern North Atlantic. [Proceeding in SciPy 2024, Tacoma, WA, USA](https://proceedings.scipy.org/articles/NVYF1037)
