### 2024-11-07 #6579 @TahiriNadia
> 4️⃣ Gagnon, J. & Tahiri, N. (2024). Ecological and Spatial Influences on the Genetics of Cumacea (Crustacea: Peracarida) in the Northern North Atlantic. [Proceeding in SciPy 2024, Tacoma, WA, USA](https://proceedings.scipy.org/articles/NVYF1037)

_aPhyloGeo_ is included as a step in our analysis, as shown in the pipeline provided in this paper.

![image](https://github.com/user-attachments/assets/be23781a-fd5a-42b6-a7cb-560fc5da42ec)
