### 2024-11-07 #6579 @TahiriNadia
Dear Editor @arfon,

We sincerely thank the reviewers, @mmore500 and @theosanderson,  for their valuable feedback, which has greatly improved the quality and clarity of our manuscript. Following their insightful comments, we have carefully addressed each requested modification. We removed self-references to enhance objectivity, and we simplified the pipeline figure. We also reviewed and ensured the precision of all decimal points in statistical tests to maintain rigorous accuracy and also the path. The licensing issue for the package was resolved by removing the bin directory, which was deemed unnecessary.

To further demonstrate the tool's value and originality, we conducted additional analyses showcasing its utility in unique ways, confirming that no existing package matches the specific functionalities of aPhyloGeo. Furthermore, we illustrated the tool’s application in peer-reviewed publications, underscoring its relevance and robustness in research contexts. Finally, we carefully revised the manuscript to reduce its length without compromising the quality of the content, ensuring a concise and focused presentation.

Sincerely,
Nadia Tahiri
