### 2024-11-09 #6579 @TahiriNadia
> Given some of the substantial clarifications made about the of role aPhyloGeo in (Li & Tahiti, 2024), I no longer feel confident in conducting this review in a good faith manner. Unfortunately, I will need to step aside as a reviewer.

The idea is not necessarily to find a link to Li and Tahiri, 2024, but to have at least one peer-reviewed paper using aPhyloGeo (Gagnon and Tahiri, 2024) that says so.

**Question**: ' is it correct to say that this analysis has not been published in a peer reviewed journal yet?'
**Answer**:  Gagnon, J. & Tahiri, N. (2024). Ecological and Spatial Influences on the Genetics of Cumacea (Crustacea: Peracarida) in the Northern North Atlantic. [Proceeding in SciPy 2024, Tacoma, WA, USA](https://proceedings.scipy.org/articles/NVYF1037)
