### 2024-11-09 #6579 @TahiriNadia
@arfon

We have implemented all the modifications requested by the reviewers, including some that we initially viewed unfavorably. They did not require multiple articles using aPhyloGeo—just one, which we have demonstrated comprehensively. We have made every possible effort, given that I have a very small team and it is challenging to find interns or students to support this project.

The primary reviewer requested adjustments including changing the license, modifying the figure, reducing text, and simplifying the visuals. The other reviewers asked us to elaborate on the scientific foundation of our results. We initially selected COVID-19 data as an illustrative example, clearly noting that the dataset was limited in size. Consequently, we embarked on an in-depth search for additional data (a challenging task) to further validate our findings, contributing to the delay. We ultimately established a collaboration with a team that provided suitable data, and we compared our results with theirs. This led to the submission, acceptance, and publication of a new article.
