### 2024-11-11 ed0133a
Add error handling for missing YAML file in Params.load_from_file method
