+++ aphylogeo/params.py
@@ -12,23 +12,23 @@ class Params:
 
     PARAMETER_KEYS = {
         "bootstrap_threshold": 0,
-        "dist_threshold": 60,
-        "window_size": 20,
-        "step_size": 100,
-        "data_names": ["ALLSKY_SFC_SW_DWN_newick", "T2M_newick", "QV2M_newick", "PRECTOTCORR_newick", "WS10M_newick"],
-        "reference_gene_dir": "./datasets/example",
-        "reference_gene_file": "sequences.fasta",
-        "file_name": "./datasets/example/geo.csv",
-        "specimen": "id",
-        "names": ["id", "ALLSKY_SFC_SW_DWN", "T2M", "PRECTOTCORR", "QV2M", "WS10M"],
+        "dist_threshold": 0,
+        "window_size": 0,
+        "step_size": 0,
+        "data_names": [],
+        "reference_gene_dir": "",
+        "reference_gene_file": "",
+        "file_name": "",
+        "specimen": "",
+        "names": [],
         "makeDebugFiles": False,
-        "bootstrap_amount": 100,
-        "alignment_method": "2",
+        "bootstrap_amount": 0,
+        "alignment_method": "0",
         "distance_method": "0",
-        "fit_method": "1",
-        "tree_type": "2",
-        "rate_similarity": 90,
-        "method_similarity": "1",
+        "fit_method": "0",
+        "tree_type": "20",
+        "rate_similarity": 0,
+        "method_similarity": "0",
     }
 
     @classmethod
