### 2024-11-22 5be0b70
Update PARAMETER_KEYS with default values
