### 2024-11-22 23ff4b7
catch maxValue - minValue division by 0
