### 2024-11-22 af1ae2f
catch 0 by division of rf_norm calculation
