### 2024-11-22 d24d039
catch division by zero by self.mem1.value
