### 2024-11-22 08cda12
catch divide by 0 by a in terminal updates
