### 2024-11-22 af82789
define bootstrap_amount
