+++ aphylogeo/params.yaml
@@ -5,7 +5,7 @@ bootstrap_threshold: 0
 dist_threshold: 60
 window_size: 200
 step_size: 100
-bootstrap_amount: 100
+bootstrap_amount: 100 #The number of bootstrap samples to generate.
 data_names: ['ALLSKY_SFC_SW_DWN_newick', 'T2M_newick', 'QV2M_newick', 'PRECTOTCORR_newick', 'WS10M_newick']
 reference_gene_dir: './datasets/example'
 reference_gene_file: 'sequences.fasta'
