### 2024-11-24 1c02f98
feat: Add intermediate step integration for the workflow
- Integrated Typer for handling CLI commands
- Implemented `climate_pipeline` and `genetic_pipeline` commands
- Set `main` function to run as default when no command is specified
- Updated `GeneticTrees` class to be iterable
- Added options for file paths in CLI commands
