+++ pyproject.toml
@@ -22,6 +22,7 @@ robinson-foulds = "^1.2"
 ete3 = "^3.1.3"
 dendropy = "^4.6.1"
 textdistance = "^4.6.0"
+typer = "^0.13.1"
 
 
 [tool.poetry.group.dev.dependencies]
