### 2024-11-24 #6579 @arfon
@TahiriNadia, thank you for your detailed responses and feedback. After reflecting on the feedback from multiple reviewers and consulting with the broader editorial team, we have decided to reject this submission. The primary concern of the JOSS editorial team is that the evidence presented here (or through past peer review of related work) does not sufficiently justify the scientific merit or utility of this software being presented here.

I will note that it is highly unusual for JOSS to reject a submission on this basis, and we do not take this decision lightly. We are always keen to find a path to publication for submissions that describe high-quality software, and we are happy to work with authors to help them improve their submission. However, in this case, we feel that the issues raised by multiple reviewers are too fundamental to be addressed in a reasonable timeframe.

I will also note that the paper, accepted as a conference submission into the SciPy proceedings, was reviewed by an astronomer (https://github.com/scipy-conference/scipy_proceedings/pull/925). While they were certainly able to provide some review feedback, we do not consider this an evaluation by a peer in your specific field.
